==> skripte_datenkorpus/5_Konsolidierung_Koerperschaft.py <==
import os
import pandas as pd

# Mapping von inkonsistenten zu konsistenten Körperschaftsnamen
koerperschaft_mapping = {
    "A.P.P.": "A.P.P. Verlag",
    "A.P.P.-Verl.": "A.P.P. Verlag",
    "A1-Verl.": "A1 Verlag",
    "AAVAA-Verl.": "AAVAA Verlag",
    "ANTHEA VERLAG": "Anthea-Verlag",
    "Anthea Verlag": "Anthea-Verlag",
    "Anthea-Verl.": "Anthea-Verlag",
    "ATHENA": "Athena-Verlag",
    "ATHENA-Verlag": "Athena-Verlag",
    "Athena": "Athena-Verlag",
    "Abl-Verl.": "Abl Verlag",
    "Acabus": "Acabus Verlag",
    "Acabus-Verl.": "Acabus Verlag",
    "acabus": "Acabus Verlag",
    "Achter Verl.": "Achter Verlag",
    "adakia Verlag UG": "adakia Verlag",
    "Adlerstein Verlag": "Adlerstein-Verlag",
    "Adlerstein-Verl.": "Adlerstein-Verlag",
    "Aagenda-Verl.": "Agenda Verlag",
    "Agenda-Verl.": "Agenda Verlag",
    "agenda Verlag": "Agenda Verlag",
    "Aisthesis-Verl.": "Aisthesis Verlag",
    "Akademie der Künste Berlin": "Akademie der Künste",
    "Alcorde-Verl.": "Alcorde Verlag",
    "Alfa-Veda Verlag": "Alfa-Veda-Verlag",
    "Alfa-Veda-Verl.": "Alfa-Veda-Verlag",
    "Alibri": "Alibri-Verl.",
    "Allitera-Verl.": "Allitera Verlag",
    "Altis-Verl.": "Altis-Verlag",
    "Ammianus-Verl.": "Ammianus-Verlag",
    "Amrûn": "Amrûn Verlag",
    "Anaconda": "Anaconda Verlag",
    "Angst&Schrecken": "Angst & Schrecken",
    "Angst&Schrecken Verlag e.K.": "Angst & Schrecken",
    "Anno-Verl.": "Anno-Verlag",
    "Appenzeller Verl.": "Appenzeller Verlag",
    "Appenzeller Verlag AG": "Appenzeller Verlag",
    "Appenzeller-Verl." : "Appenzeller Verlag",
    "AQUENSIS Verlag Pressebüro Baden-Baden GmbH": "Aquensis",
    "Aquensis Verlag Pressebüro Baden-Baden GmbH": "Aquensis",
    "Aquensis-Verl.": "Aquensis",
    "Arovell-Verl.": "Arovell Verlag",
    "Ars vivendi": "Ars Vivendi",
    "ars vivendi": "Ars Vivendi",
    "ars vivendi verlag": "Ars Vivendi",
    "ArsEd.": "Ars Edition",
    "arsEdition": "Ars Edition",
    "Art Skript Phantastik Verlag & Design": "Art Skript Phantastik",
    "Art Skript Phantastik, Verl. & Design": "Art Skript Phantastik",
    "Art- Skript-Phantastik-Verl.": "Art Skript Phantastik",
    "Art-Skript-Phantastik-Verl.": "Art Skript Phantastik",
    'art skript Phantastik "Verlag & Design"': "Art Skript Phantastik",
    "Asaro": "Asaro Verlag",
    "Asaro-Verl.": "Asaro Verlag",
    "Aschendorff": "Aschendorff Verlag",
    "Atlantik": "Atlantik Verlag",
    "Atlantis-Verl.": "Atlantis",
    "Atrium-Verl.": "Atrium Verlag",
    "Aufbau": "Aufbau Verlag",
    "Aufbau TB": "Aufbau Verlag",
    "Aufbau Taschebuch": "Aufbau Verlag",
    "Aufbau Taschenbuch": "Aufbau Verlag",
    "Aufbau-Taschenbuch": "Aufbau Verlag",
    "Aufbau-Verl.": "Aufbau Verlag",
    "Augen-Auf-Verl.": "Augen Auf Verlag",
    "August-von Goethe-Literaturverl.": "August-von-Goethe-Literaturverl.",
    "August von Goethe Literaturverlag": "August-von-Goethe-Literaturverl.",
    "Autorengemeinschaft Schreibsand": "Coburger Autorengemeinschaft Schreibsand",
    "Coburg, Schreibsand": "Coburger Autorengemeinschaft Schreibsand",
    "Autorenkreis Ruhr-Mark e.V. mit Sitz in Hagen (NRW)": "Autorenkreis Ruhr-Mark e.V.",
    "Autorenkreis Ruhr-Mark": "Autorenkfreis Ruhr-Mark e.V.",
    "BAR-Verl.": "Bar-Verlag",
    "BAR-Verlag": "Bar-Verlag",
    "BAUER-VERLAG": "Bauer-Verlag",
    "Bauer-Verlag GmbH": "Bauer-Verlag",
    "BayerVerl.": "Bayer Verlag",
    "BayerVerlag": "Bayer Verlag",
    "Bayerverl.": "Bayer Verlag",
    "Bayerischer Poeten- & Belletristik-Verl.": "Bayerischer Poeten- & Belletristik-Verlag",
    "Bayerischer Poeten- und Belletristik-Verl.": "Bayerischer Poeten- & Belletristik-Verlag",
    "Be & Be": "Be + Be",
    "Be+Be-Verlag Heiligenkreuz": "Be + Be",
    "Begedia": "Begedia Verlag",
    "Begedia-Verl.": "Begedia Verlag",
    "Benno": "St. Benno Verlag",
    "Benno-Verl.": "St. Benno Verlag",
    "St. Benno": "St. Benno Verlag",
    "St. Benno Verlag GmbH": "St. Benno Verlag",
    "benno": "St. Benno Verlag",
    "Berlin Verlag Taschenbuch": "Berlin Verlag",
    "Berlin-Verl.": "Berlin Verlag",
    "Berlin-Verl. Taschenbuch": "Berlin Verlag",
    "Berliner Festspiele / Treffen Junge Szene": "Berliner Festspiele",
    "Berner Schriftstellerinnen und Schriftsteller, BSV": "Berner Schriftstellerinnen- und Schriftsteller-Verein",
    "Schriftstellerinnen und Schriftsteller Verein, Berner": "Berner Schriftstellerinnen- und Schriftsteller-Verein",
    "Bernstein": "Bernstein-Verl.", 
    "Bertuch": "Bertuch Verlag GmbH",
    "Best-Off-Verl.": "Best-Off-Verlag",
    "Biberacher Verl.-Dr.": "Biberacher Verlagsdruckerei GmbH & Co. KG",
    "Biberacher Verlagsdr.": "Biberacher Verlagsdruckerei GmbH & Co. KG",
    "Bibliothek der Provinz": "Bibliothek der Provinz GmbH",
    "Ed. Linz, Bibliothek der Provinz": "Bibliothek der Provinz GmbH",
    "Verl. Bibliothek der Provinz": "Bibliothek der Provinz GmbH",
    "Verl. Publ. No 1, Bibliothek der Provinz": "Bibliothek der Provinz GmbH",
    "Verl. Publ. P No 1, Bibliothek der Provinz": "Bibliothek der Provinz GmbH",
    "Verlag Bibliothek der Provinz": "Bibliothek der Provinz GmbH",
    "Verlag Bibliothek der Provinz GmbH": "Bibliothek der Provinz GmbH",
    "Blaetterhaus, Verl. für Buch, Kunst & Mehr": "Blaetterhaus",
    "Blattlaus-Verl.": "Blattlaus-Verlag",
    "Blaukreuz Verlag": "Blaukreuz-Verlag",
    "Blaukreuz-Verlag Bern": "Blaukreuz-Verlag",
    "Blaulicht-Verl.": "Blaulicht-Verlag",
    "Blaulicht-Verlag, Edition Kiezpoesie": "Blaulicht-Verlag", 
    "Blue Panther Books": "Blue Panther Books",
    "Blue Panther books": "Blue Panther Books",
    "blue panther books": "Blue Panther Books",
    "BoD -Books on Demand": "BoD - Books on Demand",
    "BoD Books on Demand": "BoD - Books on Demand",
    "BoD, Books on Demand": "BoD - Books on Demand",
    "BoD, Books on demand GmbH": "BoD - Books on Demand",
    "Books on Demand GmbH &lt": "BoD - Books on Demand",
    "BoD-Books on Demand": "BoD - Books on Demand",
    "BoD": "BoD - Books on Demand",
    "BoD - Books on Demand": "BoD - Books on Demand",
    "BoD -Books on Demand": "BoD - Books on Demand",
    "BoD Books on Demand": "BoD - Books on Demand",
    "BoD GmbH": "BoD - Books on Demand",
    "BoD – Books on Demand": "BoD - Books on Demand",
    "BoD – Books on Demand – Dänemark": "BoD - Books on Demand",
    "BoD – Books on Demand – Schweden": "BoD - Books on Demand",
    "BoD, Books on Demand": "BoD - Books on Demand",
    "BoD, Books on demand GmbH": "BoD - Books on Demand",
    "BoD-Books on Demand": "BoD - Books on Demand",
    "Books on Demand": "BoD - Books on Demand",
    "Books on Demand GmbH": "BoD - Books on Demand",
    "Books on Demand Verlag GmbH": "BoD - Books on Demand",
    "Booklover": "Booklover Verlag", 
    "Booklover Verlag Nadine Kapp": "Booklover Verlag", 
    "Bouvier": "Bouvier Verlag",
    "Braumüller": "Braumüller GmbH",
    "Boldt": "Boldt-Literaturverl.",
    "BookRix": "BookRix GmbH & Co. KG",
    "Brendow": "Brendow Verlag",
    "Brendow.": "Brendow Verlag",
    "Brighton Verlag GmbH": "Brighton Verlag",
    "Brighton-Verl.": "Brighton Verlag",
    "Brockmeyer": "Brockmeyer Verlag",
    "Brot & Spiele Verlag e.U.": "Brot & Spiele Verlag",
    "Brot und Spiele": "Brot & Spiele Verlag",
    "Brot und Spiele Verlag": "Brot & Spiele Verlag",
    "Brot und Spiele Verlag e.U.": "Brot & Spiele Verlag",
    "Brot und Spiele, der Kurzgeschichtenverlag": "Brot & Spiele Verlag",
    "Brune-Mettcker": "Brune-Mettcker Druck- und Verlags-GmbH",
    "Brunnen": "Brunnen Verlag",
    "Brunnen Verlag GmbH": "Brunnen Verlag",
    "Brunnen-Verl.": "Brunnen Verlag",
    "Buch & Stabe Verlag Henning Albrecht": "Buch & Stabe Verlag",
    "Buch- & Kunstverlag Oberpfalz": "Buch- und Kunstverlag Oberpfalz",
    "Buch- und Kunst-Verl. Oberpfalz": "Buch- und Kunstverlag Oberpfalz",
    "Bucher": "Bucher Verlag GmbH",
    "Bucher GmbH & Co. Druck Verlag Netzwerk": "Bucher Verlag GmbH",
    "Buchmacherei, Scripturium Lektorat &.": "Scripturium Buchmacherei",
    "Buchschmiede": "Buchschmiede von Dataform Media GmbH",
    "Buchwerkstatt Berlin": "Buchwerkstatt",
    "Bücken & Sulzer Verlag": "Bücken & Sulzer",
    "Burg-Verl.": "Burg Verlag",
    "Burgenwelt-Verl.": "Burgenwelt Verlag",
    "Burghügel": "Burghügelverl.",
    "Burgwald-Verlag+Druck": "Burgwald-Verlag + Druck",
    "Büchse d. Pandora": "Büchse der Pandora",
    "Verlag C.H. Beck": "C.H. Beck",
    "Calvendo": "Calvendo Verlag GmbH",
    "Carl Stephenson Verlag GmbH & Co. KG": "Carl Stephenson Verlag",
    "CBK Productions": "CBK-Productions",
    "CM-Medienverl.": "CM Medienverlag",
    "CMS-Verl.-Ges.": "CMS Verlagsgesellschaft",
    "CSV, Christliche Schriftenverbreitung": "CSV, Verl. Christliche Schriftenverbreitung",
    "CW Niemeyer": "CW Niemeyer Buchverlag GmbH",
    "CW Niemeyer Buchverlage GmbH": "CW Niemeyer Buchverlag GmbH",
    "Candela-Verl.": "Candela",
    "Hanser, Carl": "Carl Hanser Verlag",
    "Carl-Auer-Verl.": "Carl-Auer Verlag GmbH",
    "Carlesso": "Carlesso Verlag",
    "Carlesso-Verl.": "Carlesso Verlag",
    "Chaotic Revelry Verl.": "Chaotic-Revelry-Verl.",
    "CharlesVerlag": "Charles Verlag",
    "Chiliverl.": "Chiliverlag",
    "Christoph-Merian-Verl.": "Christoph Merian Verlag",
    "Civitas-Imperii-Verl.": "Civitas Imperii Verlag",
    "Cleo Rivers c/o Gwendolyn Wynter Autorenimpressum": "Cleo Rivers, c/o Gwendolyn Wynter Autorenimpressum",
    "CoCon": "CoCon-Verlag GmbH",
    "Cocon": "CoCon-Verlag GmbH",
    "CoCon-Verl.": "CoCon-Verlag GmbH",
    "CoCon-Verlag": "CoCon-Verlag GmbH",
    "Cobra": "Cobra Verlag",
    "Cogitare-Verl.": "Cogitare Verlag",
    "Concepcion Seidel": "Concepcion Seidel OHG",
    "Connewitzer Verl.-Buchh. Hinke": "Connewitzer Verlagsbuchandlung",
    "Connewitzer Verlagsbuchhandlung": "Connewitzer Verlagsbuchandlung",
    "Connewitzer Vlgsbuchhdlg": "Connewitzer Verlagsbuchandlung",
    "Conte": "Conte Verlag",
    "Conte-Verl.": "Conte Verlag",
    "Conte-Verlag": "Conte Verlag",
    "Cosmos-Verl.": "Cosmos Verlag",
    "Creastro-Verlag": "Creastro Verlag",
    "CreateSpace": "CreateSpace Independent Publishing Platform",
    "CreateSpace Indipendent Publishing Platform": "CreateSpace Independent Publishing Platform",
    "Creativo Initiativgruppe für Literatur, Wissenschaft und Kunst, Wissenschaft und Kunst": "Creativo, Initiationgruppe für Literatur, Wissenschaft und KUnst",
    "Czernin": "Czernin Verlag",
    "Czernin Verlag": "Czernin Verlag",
    "Daedalus Verlag": "Daedalus Verlag",
    "Daedalus-Verl.": "Daedalus Verlag",
    "Dein-Lieblingsbuch": "Dein-Lieblingsbuch Atelier & Verlag",
    "Dein-Lieblingsbuch Atelier & Verlag": "Dein-Lieblingsbuch Atelier & Verlag",
    "Der Kleine Buch Verlag": "Der Kleine Buch-Verlag",
    "Der Kleine Buch-Verl.": "Der Kleine Buch-Verlag",
    "Der gesunde Menschenversand": "Der gesunde Menschenversand GmbH",
    "Der gesunde Menschenversand GmbH": "Der gesunde Menschenversand GmbH",
    "Die Provinz GmbH - Gemeinnützige Gesellschaft für Kulturprojekte": "Die Provinz GmbH – Kulturprojekte",
    "Die Provinz GmbH - Kulturprojekte": "Die Provinz GmbH – Kulturprojekte",
    "Die Provinz GmbH - gemeinnützige Gesellschaft für Kulturprojekte": "Die Provinz GmbH – Kulturprojekte",
    "Die Provinz GmbH – Kulturprojekte": "Die Provinz GmbH – Kulturprojekte",
    "Diogenes Theater Altstätten": "Diogenes-Theater",
    "Diogenes Theater": "Diogenes-Theater",
    "Docupoint": "Docupoint-Verlag",
    "Docupoint-Verl.": "Docupoint-Verlag",
    "docupoint GmbH": "Docupoint-Verlag",
    "docupoint  GmbH": "Docupoint-Verlag",
    "Donat": "Donat Verlag",
    "Dorante Ed.": "Dorante Edition",
    "Dorante-Ed.": "Dorante Edition",
    "Dorise": "Dorise Verlag",
    "Dorise-Verl.": "Dorise Verlag",
    "Dorise-Verlag": "Dorise Verlag",
    "dorise-Verlag": "Dorise Verlag",
    "Dortmund-Verl.": "Dortmund Verlag",
    "Dortmund-Verl.de": "Dortmund Verlag",
    "Dover Publ.": "Dover Publications, Inc.",
    "Drachenmond Verlag": "Drachenmond Verlag",
    "Drachenmond-Verl.": "Drachenmond Verlag",
    "DrachenStern Verlag": "DrachenStern Verlag",
    "Draksal Fachverlag": "Draksal Fachverlag GmbH",
    "Draupadi": "Draupadi Verlag",
    "Drava": "Drava Verlag",
    "Drava-Verl.": "Drava Verlag",
    "Drei-Birken-Verl.": "Drei Birken Verlag",
    "Dresdner Verl.": "Dresdner Verlag",
    "Dt. Literaturges.": "Deutsche Literaturgesellschaft",
    "Dt. Taschenbuch-Verl.": "Deutscher Taschenbuch Verlag",
    "Dt. Taschenbuchverl.": "Deutscher Taschenbuch Verlag",
    "Deutscher Taschenbuch-Verl.": "Deutscher Taschenbuch Verlag",
    "Dt. Lyrik-Verl. (dlv)": "Deutscher Lyrik Verlag",
    "ELYSION": "Elysion Books",
    "ELYSION-BOOKS": "Elysion Books",
    "Elysion": "Elysion Books",
    "Elysion Books": "Elysion Books",
    "Elysion-Books": "Elysion Books",
    "elysion books": "Elysion Books",
    "EPLA-Verl.": "EPLA Verlag",
    "EPLA-Verlag": "EPLA Verlag",
    "Epla-Verlag": "EPLA Verlag",
    "EPV": "EPV Elektronik-Praktiker-Verlagsgesellschaft mbH",
    "ERZDRUCK": "ERZDRUCK GmbH",
    "Ed. Märkische LebensArt": "Ed. Märkische Lebensart",
    "Ed. Paashaas": "Ed. Paashaas Verlag",
    "Ed. Paashaas-Verl.": "Ed. Paashaas Verlag",
    "Ed. PaperONE": "Ed. PaperOne",
    "Ed. Tintenfass": "Ed. Tintenfaß",
    "Edition Trailer Spot": "Edition Trailer Spot",
    "Edition Trailer Spot - der Verlag": "Edition Trailer Spot",
    "Edition Weinviertel": "Edition Weinviertel / Edition Neue Wege / Regionalverlag.at",
    "Ed. Winterwork": "Edition Winterwork",
    "Edition Winterwork": "Edition Winterwork",
    "edition Winterwork": "Edition Winterwork",
    "edition winterwork": "Edition Winterwork",
    "winterwork": "Edition Winterwork",
    "Edition Allgäu": "Edition Allgäu im Verlag HEPHAISTOS",
    "EDITION ALLGÄU im Verlag HEPHAISTOS": "Edition Allgäu im Verlag HEPHAISTOS",
    "Edition Allgäu im Verlag HEPHAISTOS": "Edition Allgäu im Verlag HEPHAISTOS",
    "Edition Aurum": "Edition Aurum Verlag",
    "Edition Blau": "Edition Blau, Rotpunktverlag",
    "Edition Blau, Belletristik im Rotpunktverlag": "Edition Blau, Rotpunktverlag",
    "Ed. Blaue Feder": "Edition Blaue Feder",
    "Ed. Fischer": "R.G. Fischer Verlag",
    "Edition Fischer": "R.G. Fischer Verlag",
    "Edition Fischer in der Verlags- und Imprintgruppe R.G. Fischer": "R.G. Fischer Verlag",
    "Edition R. G. Fischer": "R.G. Fischer Verlag",
    "Edition R.G. Fischer": "R.G. Fischer Verlag",
    "R. G. Fischer, Interbooks": "R.G. Fischer Verlag",
    "R.G. Fischer": "R.G. Fischer Verlag",
    "R.G. Fischer Interbooks": "R.G. Fischer Verlag",
    "Fischer, R. G.": "R.G. Fischer Verlag",
    "edition fischer": "R.G. Fischer Verlag",
    "edition litera im R.G. Fischer Verlag": "R.G. Fischer Verlag",
    "Ed. Nautilus": "Edition Nautilus",
    "Edition Nautilus": "Edition Nautilus",
    "Edition Nautilus GmbH": "Edition Nautilus",
    "Nautilus": "Edition Nautilus",
    "Ed. Garamond": "Edition Garamond",
    "Garamond, der Wissenschaftsverlag": "Edition Garamond",
    "Ed. Forsbach": "Edition Forsbach",
    "Ed. Blaes": "Edition Blaes",
    "Ed. Freiberg": "Edition Freiberg",
    "Eichenspinner Verlag": "Eichenspinner Verlag",
    "Eichenspinner-Verl.": "Eichenspinner Verlag",
    "Einbuch Buch- und Literaturverlag": "Einbuch Buch- und Literaturverlag",
    "Einbuch, Buch- und Literaturverl.": "Einbuch Buch- und Literaturverlag",
    "Einhorn": "Einhorn Verlag",
    "Einhorn-Verl.": "Einhorn Verlag",
    "Eire Verlag": "Eire Verlag",
    "Eire-Verl.": "Eire Verlag",
    "Elbaol Verlag": "Elbaol Verlag",
    "Elbaol Verlag Hamburg": "Elbaol Verlag",
    "Elbaol-Verl.": "Elbaol Verlag",
    "elbaol verlag hamburg": "Elbaol Verlag",
    "Elbverl.": "Elbverlag",
    "Elbverl. Rossi": "Elbverlag",
    "Elbverlag": "Elbverlag",
    "Elbverlag Magdeburg": "Elbverlag",
    "Eldur": "Eldur Verlag",
    "Eldur Verlag": "Eldur Verlag",
    "Ellert & Richter": "Ellert & Richter Verlag",
    "Ellert & Richter Verlag": "Ellert & Richter Verlag",
    "Elsinor": "Elsinor Verlag",
    "Elsinor Verlag": "Elsinor Verlag",
    "Elsinor-Verl.": "Elsinor Verlag",
    "Elster": "Elster Verlag",
    "Elster & Salis AG": "Elster Verlag",
    "Elster Verlag": "Elster Verlag",
    "Elster-Verl.": "Elster Verlag",
    "Elvea Verlag": "Elvea Verlag",
    "Elvea-Verl.": "Elvea Verlag",
    "Emmerich Book & Media": "Emmerich Books & Media",
    "Emmerich Books & Media": "Emmerich Books & Media",
    "Emons": "Emons Verlag",
    "Emons Verlag": "Emons Verlag",
    "Engeler": "Engeler Verlag",
    "Engeler-Verlag": "Engeler Verlag",
    "Engelsdorfer Verl.": "Engelsdorfer Verlag",
    "Engelsdorfer Verl.***89586": "Engelsdorfer Verlag",
    "Engelsdorfer Verlag": "Engelsdorfer Verlag",
    "Engelsdorfer-Verl.": "Engelsdorfer Verlag",
    "Engeldorfer Verl.": "Engelsdorfer Verlag",
    "Epubli, Verl.-Gruppe Holtzbrinck": "Epubli",
    "Verlag epubli GmbH": "Epubli",
    "epubli": "Epubli",
    "epubli GmbH": "Epubli",
    "epubli-Verlag": "Epubli",
    "www.epubli.de": "Epubli",
    "Eulenspiegel": "Eulenspiegel Verlag",
    "Eulenspiegel Verlag": "Eulenspiegel Verlag",
    "Eulenspiegel-Verl.": "Eulenspiegel Verlag",
    "Evang. Verl.-Anst.": "Evangelische Verlagsanstalt",
    "Evangelische Verlagsanstalt": "Evangelische Verlagsanstalt",
    "Fabuloso": "Fabuloso Verlag",
    "Fabuloso Verlag": "Fabuloso Verlag",
    "Fabuloso Verlag Gudrun Strüber": "Fabuloso Verlag",
    "Fabuloso-Verl.": "Fabuloso Verlag",
    "Fabuloso-Verl. Strüber": "Fabuloso Verlag",
    "Fabuloso-Verl. Stübner": "Fabuloso Verlag",
    "Felicitaris, Felix Marschall": "FELiCiTARiS",
    "S. FISCHER": "S. Fischer Verlag",
    "S. Fischer": "S. Fischer Verlag",
    "S. Fischer Verlag": "S. Fischer Verlag",
    "Karin Fischer Verlag": "Karin Fischer Verlag",
    "Karin Fischer Verlag GmbH": "Karin Fischer Verlag",
    "Fischer Taschenbuch": "Fischer Taschenbuch",
    "Fischer-Taschenbuch-Verl.": "Fischer Taschenbuch",
    "FISCHER Taschenbuch": "Fischer Taschenbuch",
    "Fischer": "Fischer Verlag",
    "Fischer & Fischer Medien": "Fischer Verlag",
    "Fehnland Verlag": "Fehnland Verlag",
    "Fehnland-Verlag": "Fehnland Verlag",
    "Folio Verlag": "Folio Verlag",
    "Folio-Verl.": "Folio Verlag",
    "Fontana Edizioni": "Fontana Edizioni",
    "Fontana Edizioni SA": "Fontana Edizioni",
    "Fontana edizioni": "Fontana Edizioni",
    "Fontis": "Fontis",
    "Fontis Brunnen Basel": "Fontis",
    "Frankfurter Literaturverl.": "Frankfurter Literaturverlag",
    "Frauenfelder Krimitage 2010 c/o Tourist Service Regio Frauenfeld": "Frauenfelder Krimitage",
    "Frauenfelder Krimitage &lt": "Frauenfelder Krimitage",
    "FrauenMuseum": "Frauenmuseum",
    "Free Pen Verlag": "Free Pen Verlag",
    "Free-Pen-Verl.": "Free Pen Verlag",
    "Freiburger Echo Verlag": "Freiburger Echo Verlag",
    "Freiburger-Echo-Verl. Duda": "Freiburger Echo Verlag",
    "Freya": "Freya Verlag",
    "Freya-Verl.": "Freya Verlag",
    "Frieling": "Frieling Verlag",
    "Frieling-Verlag": "Frieling Verlag",
    "Frieling-Verlag Berlin": "Frieling Verlag",
    "Förderverein für das Erich Kästner Museum": "Förderverein für das Erich-Kästner-Museum",
    "Förderverein für das Erich-Kästner-Museum": "Förderverein für das Erich-Kästner-Museum",
    "GRÄV, Verl. der GRÄGS": "GRÄV, Verlag der GRÄGS",
    "GRÄV, Verl. der Gräfelfinger Gelegenheitsschreiber": "GRÄV, Verlag der GRÄGS",
    "GRÄV, Verlag der GRÄGS": "GRÄV, Verlag der GRÄGS",
    "Gareis Marketing Kommunikation": "Gareis Marketing",
    "Gareis Marketing. Media": "Gareis Marketing",
    "Gareis Marketing.Media": "Gareis Marketing",
    "Geest": "Geest Verlag",
    "Geest-Verl.": "Geest Verlag",
    "Geest-Verlag": "Geest Verlag",
    "Geest-Verlag GmbH & Co. KG": "Geest Verlag",
    "Geiger": "Geiger Verlag",
    "Geiger-Verlag": "Geiger Verlag",
    "Geistkirch": "Geistkirch Verlag",
    "Geistkirch Verlag": "Geistkirch Verlag",
    "Geistkirch-Verl.": "Geistkirch Verlag",
    "GeistkirchVerlag": "Geistkirch Verlag",
    "Gerth Medien": "Gerth Medien",
    "Gerth Medien GmbH": "Gerth Medien",
    "GerthMedien": "Gerth Medien",
    "Girgis": "Girgis Verlag",
    "Girgis Verlag": "Girgis Verlag",
    "Glentorf": "Glentorf Verlag",
    "Glentorf-Verl.": "Glentorf Verlag",
    "Gmeiner": "Gmeiner Verlag",
    "Gmeiner-Verlag": "Gmeiner Verlag",
    "Gmeiner-Verlag GmbH": "Gmeiner Verlag",
    "Goethe-Schule-Harburg": "Goethe-Schule Harburg",
    "Goki": "Goki Verlag",
    "Goki-Verl.": "Goki Verlag",
    "Goldegg Medienakademie": "Goldegg Verlag",
    "Goldegg Verlag": "Goldegg Verlag",
    "Goldegg-Verl.": "Goldegg Verlag",
    "Gollenstein": "Gollenstein Verlag",
    "Gollenstein Verlag": "Gollenstein Verlag",
    "Grafotla? Šoltys": "Grafotlac Soltys",
    "Grafotlac Soltys": "Grafotlac Soltys",
    "Senioren-Akademie Greifswalder Bodden, Vorpommern / Gruppe Greifenfedern": '"Greifenfedern", Arbeitskreis der Senioren-Akademie Greifswalder Bodden e.V."',
    "Groh": "Groh Verlag",
    "Groh Verlag, ein Imprint der Verlagsgruppe Droemer Knaur GmbH & Co. KG": "Groh Verlag",
    "Groh Kreativteam, Groh": "Groh Verlag",
    "Groh Verlag GmbH": "Groh Verlag",
    "Gröls-Verlag, Redaktion": "Gröls Verlag",
    "Größenwahn Verlag": "Größenwahn Verlag",
    "Größenwahn-Verl.": "Größenwahn Verlag",
    "Gütersloher Verl.-Haus": "Gütersloher Verlagshaus",
    "Gütersloher Verlagshaus": "Gütersloher Verlagshaus",
    "HKL-Verl.": "HKL-Verlag",
    "HKL-Verlag": "HKL-Verlag",
    "Hai-tec Verlags- und Autorenservice": "Hai-tec Verlags- und Autorenservice",
    "Hai-tec-Verl.-und-Autorenservice": "Hai-tec Verlags- und Autorenservice",
    "Haileiht-Verlag": "Haileiht-Verlag vom Hai-tec Verlags- und Autorenservice",
    "Hamburger Lesehefte Verlag": "Hamburger Lesehefte Verlag",
    "Hamburger-Lesehefte-Verl.": "Hamburger Lesehefte Verlag",
    "Hansi' sches Dr.-  und Verl.-Haus": "Hansisches Dr.- und Verl.-Haus",
    "Hansisches Dr.- und Verl.-Haus": "Hansisches Dr.- und Verl.-Haus",
    "Hansisches Druck- und Verlagshaus": "Hansisches Dr.- und Verl.-Haus",
    "Hauptverb. des Österr. Buchhandels": "Hauptverband des Österreichischen Buchhandels",
    "Hauptverband des Österreichischen Buchhandels": "Hauptverband des Österreichischen Buchhandels",
    "Haymon": "Haymon Verlag",
    "Haymon Taschenbuch": "Haymon Verlag",
    "Haymon Verlag": "Haymon Verlag",
    "Haymon-Taschenbuch": "Haymon Verlag",
    "Haymon-Verl.": "Haymon Verlag",
    "Heimann, Stiftung": "Heimann-Stiftung, für Völkerverständigung",
    "Heimann-Stiftung, für Völkerverständigung (Hrsg.)": "Heimann-Stiftung, für Völkerverständigung",
    "Heimdall Verlag": "Heimdall Verlag",
    "Heimdall Verlag, Digital Edition": "Heimdall Verlag",
    "Heimdall-Verl. Digital Ed.": "Heimdall Verlag",
    "Helmut Lingen Verlag": "Lingen Verlag",
    "Herzsprung Verlag": "Herzsprung Verlag",
    "Herzsprung-Verlag": "Herzsprung Verlag",
    "Hierophant-Verl.": "Hierophant Verlag",
    "Hierophant-Verlag": "Hierophant Verlag",
    "Hille": "Hille Verlag",
    "Hille Verlag": "Hille Verlag",
    "Hippomonte": "Hippomonte Publishing e.K.",
    "Hippomonte Publishing e.K.": "Hippomonte Publishing e.K.",
    "Hirn Kastl & Herz": "Hirn Kastl & Herz",
    "Hirn-Kastl & Herz": "Hirn Kastl & Herz",
    "Hirnkastl & Herz": "Hirn Kastl & Herz",
    "Hirnkost": "Hirnkost KG",
    "Hirnkost KG": "Hirnkost KG",
    "Hofmann": "Hofmann-Buch",
    "Hofmann-Buch": "Hofmann-Buch",
    "Holzheimer": "Holzheimer Verlag",
    "HolzheimerVerlag": "Holzheimer Verlag",
    "Hottenstein Buchverlag": "Hottenstein Buchverlag",
    "Hottenstein-Buchverl.": "Hottenstein Buchverlag",
    "Husum": "Husum Druck- und Verlagsgesellschaft",
    "Husum Druck- und Verlagsgesellschaft": "Husum Druck- und Verlagsgesellschaft",
    "Hydre Editions": "Hydre Éditions",
    "Hydre Éd.": "Hydre Éditions",
    "Hydre Éditions": "Hydre Éditions",
    "Höll": "Höll Verlag",
    "Höll Verlag": "Höll Verlag",
    "IFB Verl. Dt. Sprache": "IFB Verlag Deutsche Sprache",
    "IFB-Verl. Dt. Sprache": "IFB Verlag Deutsche Sprache",
    "IKOS": "IKOS Verlag",
    "IKOS-Verl.": "IKOS Verlag",
    "IL-Verlag": "IL-Verlag",
    "ILV": "IL-Verlag",
    "Iatros": "Iatros Verlag",
    "Iatros Verlag": "Iatros Verlag",
    "Iatros-Verl.": "Iatros Verlag",
    "Ibera": "Ibera Verlag",
    "Ibera-Verl.": "Ibera Verlag",
    "Igel Verlag Literatur & Wissenschaft": "Igel Verlag Literatur & Wissenschaft",
    "Igel-Verl.": "Igel Verlag Literatur & Wissenschaft",
    "Igel-Verl. Literatur": "Igel Verlag Literatur & Wissenschaft",
    "ImPrint": "ImPrint Verlag",
    "ImPrint Verlag": "ImPrint Verlag",
    "ImPrint-Verl.": "ImPrint Verlag",
    "Incubus Verlag": "Incubus Verlag",
    "Incubus-Verl.": "Incubus Verlag",
    "Independent Bookworm": "Independent Bookworm",
    "Independent Bookworm Dr. Kolata": "Independent Bookworm",
    "Independent Bookworm Verlag": "Independent Bookworm",
    "Independent Bookworm, c/o Dr. Kolata": "Independent Bookworm",
    "Info Verlag GmbH": "Info Verlag",
    "Info-Verl.": "Info Verlag",
    "Input-Verl.": "Input-Verlag",
    "Input-Verlag": "Input-Verlag",
    "Insel Verlag": "Insel Verlag",
    "Insel-Verl.": "Insel Verlag",
    "Internat. Literatur- und Lyrik-Verl.": "Internationaler Literatur- und Lyrik-Verlag",
    "Internat. Literatur- und Lyrik-Verl.***5238795": "Internationaler Literatur- und Lyrik-Verlag",
    "Isensee": "Isensee Verlag",
    "Isensee Verlag": "Isensee Verlag",
    "Isensee, Florian, GmbH": "Isensee Verlag",
    "JUHR Verlag": "JUHR Verlag",
    "JUHRmade": "JUHR Verlag",
    "Jaron": "Jaron Verlag",
    "Jaron Verlag": "Jaron Verlag",
    "Jean Publishing & Distributing House": "Jean Verlag",
    "Jean Verlag": "Jean Verlag",
    "Jessica Graves c/o Klaus Zabel": "Jessica Graves",
    "Jessica Graves c/o WirFinden.es Naß und Hellie GbR": "Jessica Graves",
    "Joy Edition, Buchverlag E-Books and more": "Joy Edition",
    "Joy Edition, Buchverlag and more": "Joy Edition",
    "JuLl - Junges Literaturlabor": "JuLl – Junges Literaturlabor",
    "JuLl, Junges Literaturlabor": "JuLl – Junges Literaturlabor",
    "KBV": "KBV Verlag",
    "KBV, Edition Eyfalia": "KBV Verlag",
    "KBV-Verl.- und Medienges.": "KBV Verlag",
    "KLECKS-VERLAG": "Klecks Verlag",
    "Klecks Verlag": "Klecks Verlag",
    "Klecks-Verl.": "Klecks Verlag",
    "Klecks-Verlag": "Klecks Verlag",
    "KRAL": "Kral Verlag",
    "KRAL Verlag": "Kral Verlag",
    "Kral": "Kral Verlag",
    "Kral Verlag": "Kral Verlag",
    "Kral-Verl.": "Kral Verlag",
    "KaMeRu Verl.": "KaMeRu Verlag",
    "KaMeRu Verlag": "KaMeRu Verlag",
    "KaMeRu-Verl.": "KaMeRu Verlag",
    "Kadera Verlag": "Kadera Verlag",
    "Kadera-Verl.": "Kadera Verlag",
    "Kadera-Verlag": "Kadera Verlag",
    "Kalbhenn": "Kalbhenn, Marlies",
    "Kalbhenn, Marlies": "Kalbhenn, Marlies",
    "Karina Publishing": "Karina Verlag",
    "Karina Verlag": "Karina Verlag",
    "Karina-Publishing": "Karina Verlag",
    "Karina-Verlag": "Karina Verlag",
    "Karl Lappe Verlag": "Karl-Lappe Verlag",
    "Karl-Lappe Verlag": "Karl-Lappe Verlag",
    "Karl-May-Verl.": "Karl-May-Verlag",
    "Karl-May-Verlag": "Karl-May-Verlag",
    "- Heye Digital, Karl-May-Verlag": "Karl-May-Verlag",
    "Kastel": "Kastel, Theodor",
    "Kastel, Theodor": "Kastel, Theodor",
    "Kater": "Kater Literaturverlag",
    "Kater Literaturverlag": "Kater Literaturverlag",
    "Kaufmann": "Kaufmann Verlag",
    "Kaufmann Verlag": "Kaufmann Verlag",
    "Kein & Aber": "Kein & Aber",
    "Kein & Aber Pocket": "Kein & Aber",
    "Kelkel Verlag": "Kelkel Verlag",
    "Kelkel-Verl.": "Kelkel Verlag",
    "Kelkel-Verlag": "Kelkel Verlag",
    "Kellner": "Kellner Verlag",
    "Kellner Verlag": "Kellner Verlag",
    "Kid Verlag": "Kid Verlag",
    "Kid-Verl.": "Kid Verlag",
    "Killroy Media": "Killroy Media",
    "Killroy media": "Killroy Media",
    "Kindermund Verlag": "Kindermund Verlag",
    "Kindermund-Verl.": "Kindermund Verlag",
    "Klever": "Klever Verlag",
    "Klever Verlag": "Klever Verlag",
    "Knaur": "Knaur Verlag",
    "Knaur Taschenbuch": "Knaur Verlag",
    "Knaur-Taschenbuch-Verl.": "Knaur Verlag",
    "Kobold Books": "Kobold Books",
    "Kobold-Books": "Kobold Books",
    "Konkursbuch Verlag Claudia Gehrke": "Konkursbuch Verlag",
    "Konkursbuch-Verl. Gehrke": "Konkursbuch Verlag",
    "Kremayr & Scheriau": "Kremayr & Scheriau",
    "Kremayr Scheriau": "Kremayr & Scheriau",
    "Kulturmaschinen": "Kulturmaschinen Verlag",
    "Kulturmaschinen Verlag": "Kulturmaschinen Verlag",
    "Kulturmaschinen-Verl.": "Kulturmaschinen Verlag",
    "Kunstmann": "Kunstmann Verlag",
    "Kunstmann, A": "Kunstmann Verlag",
    "Kurznachzehn": "Kurznachzehn Verlag",
    "Kurznachzehn Verlag GmbH": "Kurznachzehn Verlag",
    "Kwalae": "Kwalae Verlag",
    "Kwalae-Verl.": "Kwalae Verlag",
    "L & L Verlag": "L & L Verlag",
    "L & L-Verlag": "L & L Verlag",
    "L&L Verlag": "L & L Verlag",
    "L&L-Verlag": "L & L Verlag",
    "Lahn-Verl.": "Lahn Verlag",
    "Lahn-Verlag in der Butzon & Bercker GmbH": "Lahn Verlag",
    "Landverl.": "Landverlag",
    "Landverlag": "Landverlag",
    "Langen Müller": "Langen Müller",
    "Langen-Müller": "Langen Müller",
    "LangenMüller": "Langen Müller",
    "Larimar Verlag": "Larimar Verlag",
    "Larimar-Verl.": "Larimar Verlag",
    "Larimar-Verlag": "Larimar Verlag",
    "Latos-Verl.": "Latos Verlag",
    "Latos-Verlag": "Latos Verlag",
    "Laumann": "Laumann Verlag",
    "Laumann-Verlag": "Laumann Verlag",
    "Lehnert": "Lehnert Verlag",
    "Lehnert Verlag": "Lehnert Verlag",
    "Lehnert, C": "Lehnert Verlag",
    "Leinpfad Verlag": "Leinpfad Verlag",
    "Leinpfad-Verl.": "Leinpfad Verlag",
    "Leipziger Literaturverl.": "Leipziger Literaturverlag",
    "Leipziger Literaturverlag": "Leipziger Literaturverlag",
    "Lektora": "Lektora Verlag",
    "Lektora GmbH": "Lektora Verlag",
    "Lenos Verlag": "Lenos Verlag",
    "Lenos Verlag **82083": "Lenos Verlag",
    "Lenos-Verl.": "Lenos Verlag",
    "Lexikus Verlag": "Lexikus Verlag",
    "Lexikus.de Verlag": "Lexikus Verlag",
    "Libelle": "Libelle Verlag",
    "Libellen-Verlag": "Libelle Verlag",
    "Lichtung Verlag GmbH": "Lichtung Verlag",
    "Lichtung-Verl.": "Lichtung Verlag",
    "Limbus Verlag": "Limbus Verlag",
    "Limbus-Verl.": "Limbus Verlag",
    "Limmat": "Limmat Verlag",
    "Limmat Verlag": "Limmat Verlag",
    "Limmat-Verl.": "Limmat Verlag",
    "Lippe Verlag": "Lippe Verlag",
    "Lippe-Verl.": "Lippe Verlag",
    "List": "List Verlag",
    "List-Taschenbuch": "List Verlag",
    "Litac": "Litac Verlag",
    "Litac GmbH": "Litac Verlag",
    "LiteraFreakPress": "LiteraFreakPress",
    "LiteraFreakPress - Ralf Preusker": "LiteraFreakPress",
    "Literareon im Utz-Verl.": "Literareon",
    "Literareon im utzverlag": "Literareon",
    "Literarische Gesellschaft (LitGes) St. Pölten": "Literarische Gesellschaft St. Pölten",
    "Literarische Gesellschaft St. Pölten": "Literarische Gesellschaft St. Pölten",
    "Literarische Gruppe Osnabrück": "Literarische Gruppe Osnabrück e.V.",
    "Literatur Quickie": "Literatur Quickie",
    "Literatur Quickie, Probsthayn": "Literatur Quickie",
    "Literatur-Quickie, Probsthayn": "Literatur Quickie",
    "Literatur-Quickie, Probsthayn & Gerlach": "Literatur Quickie",
    "Literatured. Niederösterr.": "Literaturedition Niederösterreich",
    "Literatured. Niederösterreich": "Literaturedition Niederösterreich",
    "Literaturedition Niederösterreich": "Literaturedition Niederösterreich",
    "Verlag Literaturforum Karben": "Literaturforum Karben",
    "Literaturverl. Droschl": "Literaturverlag Droschl",
    "Literaturverlag Droschl": "Literaturverlag Droschl",
    "Luchterhand": "Luchterhand Literaturverlag",
    "Luchterhand-Literaturverl.": "Luchterhand Literaturverlag",
    "Luftschacht": "Luftschacht Verlag",
    "Luftschacht Verlag": "Luftschacht Verlag",
    "Luftschacht-Verl.": "Luftschacht Verlag",
    "Luther-Verl.": "Luther Verlag",
    "Luther-Verlag": "Luther Verlag",
    "Löcker": "Löcker Verlag",
    "Löcker Verlag": "Löcker Verlag",
    "M.-G.-Schmitz- Verlag": "M.-G.-Schmitz-Verlag",
    "M.-G.-Schmitz-Verlag": "M.-G.-Schmitz-Verlag",
    "MCE Verlag": "MCE Verlag",
    "MCE Verlagsgesellschaft mbH & Co. KG (Medien Contor Elbe)": "MCE Verlag",
    "MCE Verlagsgesellschaft mbH & Co. KG": "MCE Verlag",
    "MEDU": "Medu Verlag",
    "MEDU-Verl.": "Medu Verlag",
    "Medu": "Medu Verlag",
    "Medu Verlag": "Medu Verlag",
    "Medu-Verl.": "Medu Verlag",
    "MZ Buchverlag": "MZ Buchverlag",
    "MZ-Buchverl.": "MZ Buchverlag",
    "MZ-Buchverlag": "MZ Buchverlag",
    "MYMorawa": "myMorawa",
    "myMorawa": "myMorawa",
    "myMorawa von Morawa Lesezirkel GmbH": "myMorawa",
    "myMorawa Dataform Media GmbH": "myMorawa",
    "myMorawa Lesezirkel GmbH": "myMorawa",
    "myMorawa von Dataform Media GmbH": "myMorawa",
    "myMorawa von Morawa Lesezirkel": "myMorawa",
    "Morawa Lesezirkel": "myMorawa",
    "Morawa Lesezirkel GmbH": "myMorawa",
    "Machandel Verlag": "Machandel Verlag",
    "Machandel-Verl.": "Machandel Verlag",
    "Machandel-Verlag": "Machandel Verlag",
    "Machtwort-Verl.": "Machtwortverlag",
    "Machtwortverl.": "Machtwortverlag",
    "Mackinger": "Mackingerverlag",
    "Mackingerverlag": "Mackingerverlag",
    "Make a Book": "Make a Book",
    "Make a book": "Make a Book",
    "Mandelbaum Verlag": "Mandelbaum Verlag",
    "Mandelbaum Verlag eG": "Mandelbaum Verlag",
    "Manesse": "Manesse Verlag",
    "Manesse-Verl.": "Manesse Verlag",
    "Marina Schuster": "Marina Schuster",
    "Marina Schuster c/o Block Services": "Marina Schuster",
    "Maringo": "Maringo Verlag",
    "Maringo Verl. Ingo Klopfer": "Maringo Verlag",
    "Mariposa Verlag": "Mariposa Verlag",
    "Mariposa-Verl.": "Mariposa Verlag",
    "MaroVerl.": "MaroVerlag",
    "MaroVerlag": "MaroVerlag",
    "Marterpfahl Verlag": "Marterpfahl Verlag",
    "Marterpfahl-Verl.": "Marterpfahl Verlag",
    "Marterpfahl-Verl. Happ": "Marterpfahl Verlag",
    "MASOU-Verlag": "Masou-Verlag",
    "Masou-Verlag": "Masou-Verlag",
    "Matthes & Seitz": "Matthes & Seitz Berlin",
    "Matthes & Seitz Berlin": "Matthes & Seitz Berlin",
    "Mauer-Verl.": "Mauer Verlag",
    "Mauer-Verl. Kriese": "Mauer Verlag",
    "McGinley": "McGinley Verlag",
    "McGinley-Verl.": "McGinley Verlag",
    "Mecke": "Mecke Verlag",
    "Mecke Druck und Verlag": "Mecke Verlag",
    "Mecklenburger Buchverl.": "Mecklenburger Buchverlag",
    "Mecklenburger Buchverlag": "Mecklenburger Buchverlag",
    "MedienEd. Welsch": "MedienEdition Welsch",
    "MedienEdition Welsch": "MedienEdition Welsch",
    "Memoiren-Verl.": "Memoiren-Verlag",
    "Memoiren-Verlag": "Memoiren-Verlag",
    "Memoiren-Verlag Bauschke": "Memoiren-Verlag",
    "Memoranda": "Memoranda Verlag",
    "Memoranda Verlag": "Memoranda Verlag",
    "Merkur Verlag Rinteln": "Merkur Verlag",
    "Merkur-Verl.": "Merkur Verlag",
    "Merziger Druckerei und Verl.": "Merziger-Pleban",
    "Merzinger-Pleban": "Merziger-Pleban",
    "Meyer & Meyer Verlag, Meyer &. Meyer": "Meyer & Meyer Verlag",
    "Meždunarodnyj Sojuz Nemeckoj Kul?tury": "Meždunarodnyj Sojuz Nemeckoj Kul'tury",
    "Milena Verlag": "Milena Verlag",
    "Milena-Verl.": "Milena Verlag",
    "Mirabilis Verlag": "Mirabilis Verlag",
    "Mirabilis-Verl.": "Mirabilis Verlag",
    "Mitteldeutscher Verlag": "Mitteldeutscher Verlag",
    "Mitteldt. Verl.": "Mitteldeutscher Verlag",
    "mdv Mitteldeutscher Verlag GmbH": "Mitteldeutscher Verlag",
    "mitteldeutscher verlag": "Mitteldeutscher Verlag",
    "Mitter": "Mitter Verlag",
    "Mitter-Verl.": "Mitter Verlag",
    "Mohland": "Mohland Verlag",
    "Mohland-Verl. Peters": "Mohland Verlag",
    "Mohorjeva": "Mohorjeva Hermagoras",
    "Mohorjeva Hermagoras": "Mohorjeva Hermagoras",
    "Monolith-Verl.": "Monolith Verlag",
    "Monolith-Verl. Luther-Zahn": "Monolith Verlag",
    "Monolith-Verl., Luther-Zahn": "Monolith Verlag",
    "Monsenstein & Vannerdat": "Monsenstein & Vannerdat",
    "Monsenstein und Vannerdat": "Monsenstein & Vannerdat",
    "Moritzberg Verlag": "Moritzberg Verlag",
    "Moritzberg-Verl.": "Moritzberg Verlag",
    "Morlant Verlag": "Morlant Verlag",
    "Morlant-Verlag": "Morlant Verlag",
    "Morsak Verlag": "Morsak Verlag",
    "Morsak-Verl.": "Morsak Verlag",
    "Morsakverlag": "Morsak Verlag",
    "Muc Verlag": "Muc Verlag",
    "Muc-Verl.": "Muc Verlag",
    "MylinFalkläm": "Mylinfalklaem",
    "Mylinfalklaem": "Mylinfalklaem",
    "Männerschwarm Verlag": "Männerschwarm Verlag",
    "Männerschwarm-Verl.": "Männerschwarm Verlag",
    "Mönnig": "Mönnig Verlag",
    "Mönnig-Verlag": "Mönnig Verlag",
    "Münchner Turmschreiber e.V.": "Die Münchner Turmschreiber e.V.",
    "Münster Verlag": "Münster Verlag",
    "Münsterverl.": "Münster Verlag",
    "NORA": "NORA Verlag",
    "NORA, Novitäten & Raritäten": "NORA Verlag",
    "Nora": "NORA Verlag",
    "Nora, Novitäten & Raritäten": "NORA Verlag",
    "NOTschr.-Verl.": "NOTschriften Verlag",
    "NOTschriften": "NOTschriften Verlag",
    "NOTschriften Verlag": "NOTschriften Verlag",
    "NOTschriften-Verlag": "NOTschriften Verlag",
    "Notschr.-Verl.": "NOTschriften Verlag",
    "Notschriften": "NOTschriften Verlag",
    "Notschriften Verlag": "NOTschriften Verlag",
    "Notschriften-Verl.": "NOTschriften Verlag",
    "Notschriften-Verlag": "NOTschriften Verlag",
    "Natalie Rabengut": "Natalie Rabengut",
    "Natalie Rabengut c/o Black Umbrella Publishing": "Natalie Rabengut",
    "Naumann": "Naumann & Göbel",
    "Naumann & Göbel Verlagsgesellschaft mbH": "Naumann & Göbel",
    "Nee Popelubu": "Nee Popelubu",
    "Nee Popelubu c/o Wagner": "Nee Popelubu",
    "Net-Verl.": "Net-Verlag",
    "net-Verl.": "Net-Verlag",
    "net-Verlag": "Net-Verlag",
    "Neukirchener": "Neukirchener Verlag",
    "Neukirchener Verlag": "Neukirchener Verlag",
    "Neukirchener Verlagsgesellschaft": "Neukirchener Verlag",
    "NEXX Verlag GmbH": "Nexx Verlag",
    "Niederland-Verl. Liessel": "Niederland Verlag",
    "Niederlandverl. Liessel": "Niederland Verlag",
    "Nordpark": "Nordpark Verlag",
    "NordPark": "Nordpark Verlag",
    "Nova MD": "Nova MD",
    "Nova MD GmbH": "Nova MD",
    "Novum Eco": "novum Verlag",
    "Novum Pocket": "novum Verlag",
    "Novum Pro": "novum Verlag",
    "Novum pro": "novum Verlag",
    "Novum-Verl.": "novum Verlag",
    "novum Verlag in der Verlags- und Medienhaus WSB": "novum Verlag",
    "novum eco Verlag": "novum Verlag",
    "novum pocket": "novum Verlag",
    "novum pocket Verlag": "novum Verlag",
    "novum premium": "novum Verlag",
    "novum premium ein Imprint von novum publishing": "novum Verlag",
    "novum pro": "novum Verlag",
    "novum pro Verlag": "novum Verlag",
    "Null Papier Frisch": "Null Papier Verlag",
    "Null Papier Verlag": "Null Papier Verlag",
    "OSL": "OSL",
    "OSL Verlag": "OSL",
    "OSL-Verl.": "OSL",
    "Oberlausitzer Verl.": "Oberlausitzer Verlag",
    "Oberlausitzer Verl. Nürnberger": "Oberlausitzer Verlag",
    "Oberlausitzer Verlag": "Oberlausitzer Verlag",
    "Oberschlesisches Eichendorff-Kultur- und Begegnungszentrum": "Oberschlesisches Eichendorff-Kultur- und Begegnungszentrum",
    "Oberschlesisches Joseph-von-Eichendorff-Kultur- und Begegnungszentrum": "Oberschlesisches Eichendorff-Kultur- und Begegnungszentrum",
    "Odenwald-Verl.": "Odenwald-Verlag",
    "Odenwald-Verlag": "Odenwald-Verlag",
    "Offizin": "Offizin Verlag",
    "Offizin Verlag GmbH": "Offizin Verlag",
    "Ohetaler Verl.": "Ohetaler Verlag",
    "Ohetaler Verl., Heimatverein d'Ohetaler Riedlhütte": "Ohetaler Verlag",
    "Ohetaler Verlag": "Ohetaler Verlag",
    "Omnino": "Omnino Verlag",
    "Omnino Verlag": "Omnino Verlag",
    "Omnino.": "Omnino Verlag",
    "Opus Magnum": "Opus Magnum",
    "Opus magnum": "Opus Magnum",
    "OsirisDr.": "OsirisDruck",
    "OsirisDruck": "OsirisDruck",
    "Ostfriesland Verlag": "Ostfriesland Verlag",
    "Ostfriesland Verlag - SKN": "Ostfriesland Verlag",
    "Otto-Johann Verlag": "Otto-Johann-Verlag",
    "Otto-Johann-Verl.": "Otto-Johann-Verlag",
    "PMLakeman Verlag": "PMLakeman-Verlag",
    "PMLakeman-Verlag": "PMLakeman-Verlag",
    "PROOF Druck- und Medienproduktion": "PROOF Verlag",
    "PROOF Verlag": "PROOF Verlag",
    "PROOF Verlag Erfurt": "PROOF Verlag",
    "Pandmonium Verlag": "Pandämonium-Verlag",
    "Pandämonium Verlag": "Pandämonium-Verlag",
    "Pandämonium-Verl.": "Pandämonium-Verlag",
    "Panga-Misi Verlag": "Panga-Misi Verlag",
    "Pangai-Misi-Verl.": "Panga-Misi Verlag",
    "Papierflieger": "Papierflieger",
    "Papierflieger-Verl.": "Papierflieger",
    "Papierfresserchens MTM-Verl.": "Papierfresserchens MTM-Verlag",
    "Papierfresserchens MTM-Verlag": "Papierfresserchens MTM-Verlag",
    "Papierfresserchens MTM-Verlag + Herzsprung-Verlag": "Papierfresserchens MTM-Verlag",
    "Papierfresserchens MTM-Verlag GbR": "Papierfresserchens MTM-Verlag",
    "Paramon": "Paramon Verlag",
    "Paramon Verlag": "Paramon Verlag",
    "Parzeller": "Parzellers Buchverlag",
    "Parzellers Buchverlag und Werbemittel GmbH & Co. KG": "Parzellers Buchverlag",
    "Passage-Verl.": "Passage-Verlag",
    "Passage-Verlag": "Passage-Verlag",
    "Passagen Verlag": "Passage-Verlag",
    "Passagen-Verl.": "Passage-Verlag",
    "Passanten Verlag": "Passanten Verlag",
    "Passanten-Verl.": "Passanten Verlag",
    "Patmos Verlag": "Patmos Verlag",
    "Patmos-Verl.": "Patmos Verlag",
    "Pattloch": "Pattloch Verlag",
    "Pattloch Geschenkbuch": "Pattloch Verlag",
    "Pattloch Verlag": "Pattloch Verlag",
    "Paulus": "Paulusverlag",
    "Paulusverl.": "Paulusverlag",
    "Pegasus-Verein für Kreatives Schreiben": "Pegasus Verein für Kreatives Schreiben e.V.",
    "Penguin": "Penguin Verlag",
    "Penguin Verlag": "Penguin Verlag",
    "Periplaneta": "Periplaneta",
    "Periplaneta Berlin, Edition Subkultur": "Periplaneta",
    "Persimplex Verl.": "Persimplex-Verlag",
    "Persimplex-Verl.": "Persimplex-Verlag",
    "Persimplex-Verl.-Gruppe": "Persimplex-Verlag",
    "Pflegedienst, Rückenwind": "Rückenwind Pflegedienst",
    "Ph. C.W. Schmidt": "VERLAG PH. C.W. SCHMIDT",
    "Picus": "Picus Verlag",
    "Picus Verlag": "Picus Verlag",
    "Picus-Verl.": "Picus Verlag",
    "Pigmentar": "Pigmentar Verlag",
    "Pigmentar GmbH": "Pigmentar Verlag",
    "Pigmentar GmbH/Verlag": "Pigmentar Verlag",
    "Pilger-Verl.": "Pilgerverlag",
    "Pilgerverlag": "Pilgerverlag",
    "Pilum Literatur Verlag": "Pilum Literatur Verlag",
    "Pilum Verlag": "Pilum Literatur Verlag",
    "Pilum-Literatur-Verl.": "Pilum Literatur Verlag",
    "Piper Verlag": "Piper",
    "PoCul": "PoCul-Verlag",
    "PoCul-Verl. für Politik und Cultur": "PoCul-Verlag",
    "PoCul-Verlag": "PoCul-Verlag",
    "Pop": "Pop Verlag",
    "Pop Verlag": "Pop Verlag",
    "Pop Verlag  Literatur": "Pop Verlag",
    "Pop Verlag Literatur": "Pop Verlag",
    "POP-Verlag": "Pop Verlag",
    "Printsystem Medienverlag": "Printsystem Medienverlag",
    "Printsystem-Medienverl.": "Printsystem Medienverlag",
    "Printsystem-Medienverlag": "Printsystem Medienverlag",
    "Pro BUSINESS": "Pro BUSINESS",
    "Pro BUSINESS GmbH": "Pro BUSINESS",
    "Pro Business": "Pro BUSINESS",
    "Pro Business digital": "Pro BUSINESS",
    "Pro-Business-Verl.": "Pro BUSINESS",
    "Pro Libro": "Pro Libro",
    "Pro Libro GmbH": "Pro Libro",
    "Pro Libro Luzern": "Pro Libro",
    "Projekt-Verl.": "Projekt Verlag",
    "projektverlag": "Projekt Verlag",
    "Projekte Verlag Hahn": "Projekt Verlag",
    "Projekte-Verl": "Projekt Verlag",
    "Projekte-Verl.": "Projekt Verlag",
    "Projekte-Verl. Cornelius": "Projekt Verlag",
    "Projekte-Verl. Cornelius***5105763": "Projekt Verlag",
    "Projekte-Verl. Cornelsen": "Projekt Verlag",
    "Projekte-Verlag Cornelius": "Projekt Verlag",
    "Projektverl.": "Projekt Verlag",
    "Projektverlag": "Projekt Verlag",
    "ProMÖLLTAL-Initiative für Bildung, Kultur, Wirtschaft und Tourismus": "ProMÖLLTAL - Initiative für Bildung, Kultur, Wirtschaft und Tourismus",
    "ProMÖLLTAL – Initiative für Bildung, Kultur, Wirtschaft und Tourismus": "ProMÖLLTAL - Initiative für Bildung, Kultur, Wirtschaft und Tourismus",
    "Präsenz": "Präsenz Medien & Verlag",
    "Präsenz Medien & Verlag": "Präsenz Medien & Verlag",
    "Public book media Verlag": "Public book media Verlag",
    "Public-Book-Media-Verl.": "Public book media Verlag",
    "Queich Verlag": "Queich Verlag",
    "Queich-Verl.": "Queich Verlag",
    "Quickborn": "Quickborn-Verlag",
    "Quickborn-Verl.": "Quickborn-Verlag",
    "Quickborn-Verlag": "Quickborn-Verlag",
    "Quintus": "Quintus-Verlag",
    "Quintus-Verlag": "Quintus-Verlag",
    "Rheingauer Mundartverein": "Rheingauer Mundartverein e. V.",
    "Rowohlt Taschenbuch": "Rowohlt Taschenbuch Verlag",
    "Rowohlt Taschenbuch VErlag": "Rowohlt Taschenbuch Verlag",
    "Rowohlt Taschenbuch Verlag": "Rowohlt Taschenbuch Verlag",
    "Rowohlt-Taschenbuch-Verl.": "Rowohlt Taschenbuch Verlag",
    "Radius": "Radius Verlag",
    "Radius Verlag": "Radius Verlag",
    "Rautenberg": "Rautenberg",
    "Rautenberg Media": "Rautenberg",
    "Rautenberg im Verl.-Haus Würzburg": "Rautenberg",
    "Re-Di-Roma-Verl.": "Rediroma Verlag",
    "Rediroma Verlag": "Rediroma Verlag",
    "Rediroma-Verl.": "Rediroma Verlag",
    "Rediroma-Verlag": "Rediroma Verlag",
    "Reader's Digest": "Reader's Digest",
    "Reader's Digest Deutschland, Schweiz, Österreich": "Reader's Digest",
    "Reclam": "Reclam",
    "Reclam, Philipp": "Philipp Reclam jun. Verlag",
    "Redieck & Schade": "Redieck & Schade",
    "Redieck & Schade GbR": "Redieck & Schade",
    "Regia": "Regia",
    "Regia-Co-Work": "Regia",
    "Regia-Verl.": "Regia",
    "Reimo-Verl.": "Reimo-Verlag",
    "Reimo-Verlag": "Reimo-Verlag",
    "Reinecke & Voß": "Reinecke & Voß",
    "Reinecke & Voß c/o Betram Reinecke": "Reinecke & Voß",
    "Reisebuch Verlag": "Reisebuch Verlag",
    "Reisebuchverl. Kastanienhof": "Reisebuch Verlag",
    "Renaissance-Verl.": "Renaissance-Verl. Marburg",
    "Renaissance-Verl. Marburg": "Renaissance-Verl. Marburg",
    "Residenz": "Residenz Verlag",
    "Residenz Verlag": "Residenz Verlag",
    "Residenz-Verl.": "Residenz Verlag",
    "Resistenz": "Resistenz-Verl.",
    "Resistenz-Verl.": "Resistenz-Verl.",
    "Rhein-Mosel-Verl.": "Rhein-Mosel-Verlag",
    "Rhein-Mosel-Verlag": "Rhein-Mosel-Verlag",
    "Rimbaud": "Rimbaud Verlag",
    "Rimbaud Verlag": "Rimbaud Verlag",
    "Romeon Verlag": "Romeon Verlag",
    "Romeon-Verlag": "Romeon Verlag",
    "Rosamontis-Verl": "Rosamontis-Verl.",
    "Rotabene Data Service": "Rotabene Medienhaus",
    "Rotabene Medienhaus": "Rotabene Medienhaus",
    "Rotabene-Medienhaus": "Rotabene Medienhaus",
    "Verlag Rotabene Medienhaus": "Rotabene Medienhaus",
    "Rotpunktverl.": "Rotpunktverlag",
    "Rotpunktverlag": "Rotpunktverlag",
    "Rotten Verlag": "Rotten Verlag",
    "Rotten Verlags AG": "Rotten Verlag",
    "Rotten-Verl.": "Rotten Verlag",
    "Ruhrkrimi-Verlag": "Ruhrkrimi-Verlag",
    "Ruhrkrimi-Verlag Uwe Wittenfeld": "Ruhrkrimi-Verlag",
    "S. Sell Heimat-Verlag": "S. Sell Heimat-Verlag",
    "S. Sell Heimat-Verlag Altenburg": "S. Sell Heimat-Verlag",
    "SCM": "SCM",
    "SCM Coll.": "SCM",
    "SCM Hänssler": "SCM",
    "SCM R. Brockhaus": "SCM",
    "SJW": "SJW",
    "SJW Schweizerisches Jugendschriftenwerk": "SJW",
    "SJW, Schweizerisches Jugendschriftenwerk": "SJW",
    "SWB-Media-Publ.": "SWB-Verl.",
    "SWB-Verl.": "SWB-Verl.",
    "SWB.-Verl.": "SWB-Verl.",
    "Salier": "Salier Verlag",
    "Salier Verlag": "Salier Verlag",
    "Salon-Literatur-Verl.": "Salon-LiteraturVerl.",
    "Salon-LiteraturVerl.": "Salon-LiteraturVerl.",
    "Salsa Verlag": "Salsa Verlag",
    "Salsa Verlag Britta Voß": "Salsa Verlag",
    "Samenkorn": "Samenkorn e.V.",
    "Samenkorn e.V.": "Samenkorn e.V.",
    "Sarah Baines c/o madera GmbH & Co.": "Sarah Baines c/o madera GmbH & Co. KG",
    "Sarah Baines c/o madera GmbH & Co. KG": "Sarah Baines c/o madera GmbH & Co. KG",
    "Sarturia Verlag": "Sarturia Verlag e.K. Autoren Service",
    "Sarturia Verlag e.K. Autoren Service": "Sarturia Verlag e.K. Autoren Service",
    "Sarturia Verlag e.K. Autoren-Service": "Sarturia Verlag e.K. Autoren Service",
    "Sarturia e.K. Verlag Autoren Service": "Sarturia Verlag e.K. Autoren Service",
    "Sarturia-Autoren-Community": "Sarturia Verlag e.K. Autoren Service",
    "Sarturia-Verl.": "Sarturia Verlag e.K. Autoren Service",
    "Sarturia-Verlag e.K. Autoren Service": "Sarturia Verlag e.K. Autoren Service",
    "Satyr Verlag": "Satyr-Verl.",
    "Satyr-Verl.": "Satyr-Verl.",
    "Schardt": "Schardt Verlag",
    "Schardt Verlag": "Schardt Verlag",
    "Schardt, M": "Schardt Verlag",
    "Schardt Verlag e.K.": "Schardt Verlag",
    "Schibri-Verl.": "Schibri-Verlag",
    "Schibri-Verlag": "Schibri-Verlag",
    "Schiller": "Schiller Verlag",
    "Schiller Verlag": "Schiller Verlag",
    "Schillinger": "Schillinger Verlag",
    "Schillinger Verlag": "Schillinger Verlag",
    "Schneider-Verl. Hohengehren": "Schneider Hohengehren",
    "Schneider Hohengehren": "Schneider Hohengehren",
    "Scholastika Verlag": "Scholastika-Verlag",
    "Scholastika-Verlag": "Scholastika-Verlag",
    "Schreiblust-Verl.": "Schreiblust-Verlag",
    "Schreiblust-Verlag": "Schreiblust-Verlag",
    "Schreibwerk Ost bei Edition Howeg": "Schreibwerkstatt Ost bei Edition Howeg",
    "Schreibwerkstatt Ost bei Edition Howeg": "Schreibwerkstatt Ost bei Edition Howeg",
    "Schröder": "Schröder Verlag",
    "Schröder Verlag": "Schröder Verlag",
    "Schweitzerhaus Verlag": "Schweitzerhaus-Verl.",
    "Schweitzerhaus-Verl.": "Schweitzerhaus-Verl.",
    "Schweizer Literaturges.": "Schweizer Literaturgesellschaft",
    "Schweizer Literaturgesellschaft": "Schweizer Literaturgesellschaft",
    "Schöffling": "Schöffling & Co.",
    "Schöffling & Co.": "Schöffling & Co.",
    "Scribaria &lt": "scribaria",
    "Scribo Verlagsgesellschaft": "Scribo Verlag",
    "Secession Verlag für Literatur": "Secession, Verl. für Literatur",
    "Secession, Verl. für Literatur": "Secession, Verl. für Literatur",
    "Seniorenschreibtisch der Gemeinde Puchheim": "Seniorenschreibtisch der Stadt Puchheim",
    "Seniorenschreibtisch der Stadt Puchheim": "Seniorenschreibtisch der Stadt Puchheim",
    "Septime Verlag": "Septime Verlag",
    "Septime-Verl.": "Septime Verlag",
    "Servus": "Servus Verlag",
    "Servus Verlag": "Servus Verlag",
    "Shaker": "Shaker Verlag",
    "Shaker Media": "Shaker Verlag",
    "Shaker Media GmbH": "Shaker Verlag",
    "Shaker Verlag": "Shaker Verlag",
    "Sieben Verlag": "Sieben Verlag",
    "Sieben-Verl.": "Sieben Verlag",
    "Sieben-Verlag": "Sieben Verlag",
    "Silberburg": "Silberburg Verlag",
    "Silberburg Verlag": "Silberburg Verlag",
    "Silberburg-Verl.": "Silberburg Verlag",
    "Silberburg-Verlag": "Silberburg Verlag",
    "Silke Heinig": "Silke Heinig",
    "Silke Heinig, Literatur Treff Grünau": "Silke Heinig",
    "SingLiesel": "SingLiesel Verlag",
    "SingLiesel GmbH": "SingLiesel Verlag",
    "SingLiesel Verlag": "SingLiesel Verlag",
    "Singliesel GmbH": "SingLiesel Verlag",
    "Skarabaeus": "Skarabäus",
    "Skarabäus": "Skarabäus",
    "Societäts Verlag": "Societäts-Verlag",
    "Societäts-Verl.": "Societäts-Verlag",
    "Societäts-Verlag": "Societäts-Verlag",
    "Solibro": "Solibro-Verl.",
    "Solibro-Verl.": "Solibro-Verl.",
    "Songdog Verlag": "Songdog-Verl.",
    "Songdog-Verl.": "Songdog-Verl.",
    "SoralPRO Verlag": "SoralPro-Verl.",
    "SoralPro Verlag": "SoralPro-Verl.",
    "SoralPro-Verl.": "SoralPro-Verl.",
    "Space Net": "SpaceNet Verlag",
    "SpaceNet AG": "SpaceNet Verlag",
    "SpaceNet Verlag": "SpaceNet Verlag",
    "Spaß am Lesen": "Spaß am Lesen Verlag",
    "Spaß am Lesen Verlag": "Spaß am Lesen Verlag",
    "Spaß am Lesen Verlag GmbH, Spaß am Lesen Verlag": "Spaß am Lesen Verlag",
    "Sperling Verlag": "Sperling-Verlag",
    "Sperling-Verl.": "Sperling-Verlag",
    "Sperling-Verlag": "Sperling-Verlag",
    "Spica": "Spica-Verl.",
    "Spica Verlag GmbH": "Spica-Verl.",
    "Spica-Verl.": "Spica-Verl.",
    "Spielberg": "Spielberg Verlag",
    "Spielberg Verlag": "Spielberg Verlag",
    "Spielberg-Verl.": "Spielberg Verlag",
    "Spurbuchverl.": "Spurbuchverlag",
    "Spurbuchverlag": "Spurbuchverlag",
    "Stephenson": "Stephenson, C",
    "Stephenson, C": "Stephenson, C",
    "Sternensand Verlag": "Sternensand Verlag",
    "Sternensand Verlag GmbH": "Sternensand Verlag",
    "Storyhouse Verlag": "Storyhouse-Verl.",
    "Storyhouse-Verl.": "Storyhouse-Verl.",
    "Styria Premium": "Styria Verlag",
    "Styria Regional": "Styria Verlag",
    "Styria Verlag": "Styria Verlag",
    "Suhrkamp": "Suhrkamp Verlag",
    "Suhrkamp Verlag": "Suhrkamp Verlag",
    "Sujet": "Sujet Verlag",
    "Sujet Verlag": "Sujet Verlag",
    "Sujet-Verl.": "Sujet Verlag",
    "SüdOst Service GmbH": "SüdOst Verlag",
    "SüdOst-Verl.": "SüdOst Verlag",
    "Südostschweiz Presse und Print": "Südostschweiz-Buchverl.",
    "Südostschweiz-Buchverl.": "Südostschweiz-Buchverl.",
    "TEXT/RAHMEN": "TEXT/RAHMEN e.U.",
    "TEXT/RAHMEN e.U.": "TEXT/RAHMEN e.U.",
    "Text/Rahmen": "TEXT/RAHMEN e.U.",
    "Text/Rahmen Taschen": "TEXT/RAHMEN e.U.",
    "THK Verlag": "THK-Verlag UG",
    "THK-Verlag UG": "THK-Verlag UG",
    "TRIGA - Der Verlag": "TRIGA - Der Verlag",
    "TRIGA - Der Verlag UG**5107566": "TRIGA - Der Verlag",
    "TRIGA - der Verl.": "TRIGA - Der Verlag",
    "TRIGA Der Verlag": "TRIGA - Der Verlag",
    "TRIGA Der Verlag Gerlinde Heß": "TRIGA - Der Verlag",
    "TWENTYSIX": "TWENTYSIX",
    "TWENTYSIX CRIME": "TWENTYSIX",
    "TWENTYSIX EPIC": "TWENTYSIX",
    "TWENTYSIX LOVE": "TWENTYSIX",
    "Tauber-Verl.": "Tauberverlag",
    "Tauberverl.": "Tauberverlag",
    "Tauberverlag": "Tauberverlag",
    "Telescope Verlag": "Telescope-Verl.",
    "Telescope-Verl.": "Telescope-Verl.",
    "Tierbuchverl. Hohe": "Tierbuchverlag Irene Hohe",
    "Tierbuchverlag Irene Hohe": "Tierbuchverlag Irene Hohe",
    "TAK - Tiroler Autorinnen-und-Autoren-Kooperative": "TAK - Tiroler Autorinnen und Autoren Kooperative",
    "Tiroler Autorinnen und Autoren Kooperative": "TAK - Tiroler Autorinnen und Autoren Kooperative",
    "ToMa": "ToMa",
    "ToMa, Papierfresserchens MTM-Verl.": "ToMa",
    "ToMa-Ed.": "ToMa",
    "TösterKultur e.V. (Hg.), TösterKultur e.V.": "TösterKultur e.V.",
    "Treibgut Verlag": "Treibgut-Verl.",
    "Treibgut-Verl.": "Treibgut-Verl.",
    "Trotz Verlag": "Trotz-Verl.",
    "Trotz-Verl.": "Trotz-Verl.",
    "Turmschreiber Verlag": "Turmschreiber-Verl.",
    "Turmschreiber Verlag Ingwert Paulsen jr.": "Turmschreiber-Verl.",
    "Turmschreiber-Verl.": "Turmschreiber-Verl.",
    "Twilight-Line": "Twilight-Line Verlag",
    "Twilight-Line Medien GbR": "Twilight-Line Verlag",
    "Twilight-Line Verlag GbR": "Twilight-Line Verlag",
    "Twilight-Line medien GbR": "Twilight-Line Verlag",
    "Twilight-Line-Verl.": "Twilight-Line Verlag",
    "Typostudio Schumacher-Gebler": "Typostudio Schumacher-Gebler",
    "Typostudio SchumacherGebler": "Typostudio Schumacher-Gebler",
    "U-Line UG": "U-Line-Verl.",
    "U-Line-Verl.": "U-Line-Verl.",
    "U. Misslitz": "U. Mißlitz",
    "U. Mißlitz": "U. Mißlitz",
    "UND-Verl.": "UND-Verlag",
    "UND-Verlag": "UND-Verlag",
    "Ubooks": "Ubooks-Verl.",
    "Ubooks-Verl.": "Ubooks-Verl.",
    "Ulisses Medien und Spiel Distribution GmbH": "Ulisses Spiele",
    "Ulisses Spiele": "Ulisses Spiele",
    "Ulisses Spiele GmbH": "Ulisses Spiele",
    "Ullstein": "Ullstein Verlag",
    "Ullstein Taschenbuch Verlag": "Ullstein Verlag",
    "Ullstein Verlag": "Ullstein Verlag",
    "Ullstein extra": "Ullstein Verlag",
    "UniScripta": "UniScripta-Verl.",
    "UniScripta Verlag": "UniScripta-Verl.",
    "UniScripta-Verl.": "UniScripta-Verl.",
    "Unionsverl.": "Unionsverlag",
    "Unionsverlag": "Unionsverlag",
    "United P. C.": "United p.c.",
    "United P.C.": "United p.c.",
    "United p.c.": "United p.c.",
    "united p.c.": "United p.c.",
    "Herbert Utz Verlag": "utzverlag",
    "VAS": "VAS - Verlag für Akademische Schriften",
    "VAS - Verlag für Akademische Schriften": "VAS - Verlag für Akademische Schriften",
    "VHV Literatur & Kultur Verlag": "VHV, Literatur & Kultur Verlag",
    "VHV, Literatur & Kultur Verlag": "VHV, Literatur & Kultur Verlag",
    "Vantage Point World-Verl.": "Vantage-Point-World-Verl.",
    "Vantage-Point-World-Verl.": "Vantage-Point-World-Verl.",
    "Vehling": "Vehling Verlag",
    "Vehling Medienservice & Verlag GmbH": "Vehling Verlag",
    "Vehling Verlag": "Vehling Verlag",
    "VentVerl.": "VentVerl.",
    "VentVerl. Vent-Schmidt": "VentVerl.",
    "Ventil": "Ventil-Verl.",
    "Ventil-Verl.": "Ventil-Verl.",
    "Ventura Verlag": "Ventura-Verl.",
    "Ventura-Verl.": "Ventura-Verl.",
    "Verbrecher Verlag": "Verbrecher-Verl.",
    "Verbrecher-Verl.": "Verbrecher-Verl.",
    "Verein Alltag Verlag": "Verein Alltag-Verl.",
    "Verein Alltag-Verl.": "Verein Alltag-Verl.",
    "Verein zur Förderung des Literaturschaffens und des Kulturellen Austauschs Zwischen Deutschen und Polen e.V.": "Verein zur Förderung des Literaturschaffens zwischen Deutschen und Polen",
    "Verein zur Förderung des Literaturschaffens und des Kulturellen Austauschs zwischen Deutschen und Polen": "Verein zur Förderung des Literaturschaffens zwischen Deutschen und Polen",
    "Verein zur Förderung des Literaturschaffens und des Kulturellen Austauschs zwischen Deutschen und Polen e. V.": "Verein zur Förderung des Literaturschaffens zwischen Deutschen und Polen",
    "Verein zur Förderung des Literaturschaffens und des Kulturellen Austauschs zwischen Deutschen und Polen e.V.": "Verein zur Förderung des Literaturschaffens zwischen Deutschen und Polen",
    "Verein zur Förderung des Literaturschaffens und des kulturellen Austauschs zwischen Deutschen und Polen e.V.": "Verein zur Förderung des Literaturschaffens zwischen Deutschen und Polen",
    "Galiani Berlin": "Galiani Berlin",
    "Verl. Galiani": "Galiani Berlin",
    "Verlag, ChaWila": "Chawila",
    "Verlag Expeditionen": "Verlag Expeditionen",
    "Verlag Expeditionen, Edition Hamburger Autorenvereinigung": "Verlag Expeditionen",
    "Verlag Schönrock & Heikamp": "Verlag Schönrock & Heikamp",
    "Verlag Schönrock & Heikamp GbR": "Verlag Schönrock & Heikamp",
    "Verlag Vitolibro": "Vitolibro",
    "Vitolibro": "Vitolibro",
    "vitolibro": "Vitolibro",
    "vitolibro Vito von Eichborn": "Vitolibro",
    "Verlag Voland & Quist": "Voland & Quist",
    "Voland & Quist": "Voland & Quist",
    "Verlag am Eschbach": "Verlag am Eschbach",
    "Verlag am Eschbach der Schwabenverlag AG": "Verlag am Eschbach",
    "Verlag am Eschbach GmbH": "Verlag am Eschbach",
    "Verlag des Institut Drachenhaus": "Drachenhaus Verlag",
    "Verlag des Instituts Drachenhaus": "Drachenhaus Verlag",
    "Verlag für Moderne Phantastik Gehrke": "Verlag für Moderne Phantastik Gehrke",
    "Verlag für Moderne Phantastik Gehrke VPMG": "Verlag für Moderne Phantastik Gehrke",
    "ViCON-Verl.": "ViCON-Verlag",
    "ViCON-Verlag": "ViCON-Verlag",
    "Vidal": "Vidal Verlag",
    "Vidal Verlag": "Vidal Verlag",
    "Vidal Verlag Winterthur": "Vidal Verlag",
    "Vincent Theodor Thomas": "Vincent Theodor Thomas",
    "Vincent Theodor Thomas Selbstverlag": "Vincent Theodor Thomas",
    "Vindobona": "Vindobona Verlag",
    "Vindobona Verlag": "Vindobona Verlag",
    "Vindobona Verlag in der Verlags- und Medienhaus WSB GmbH": "Vindobona Verlag",
    "Vindobona-Verl": "Vindobona Verlag",
    "Vindobona-Verl.": "Vindobona Verlag",
    "Viola Plötz alias Ava Innings / Violet Truelove": "Viola Plötz",
    "Viola Plötz alias Ava Innings/Violet Truelove": "Viola Plötz",
    "WFB": "WFB-Verlagsgruppe",
    "WFB-Verl.-Gruppe": "WFB-Verlagsgruppe",
    "WOLL Verlag": "WOLL Verlag",
    "WOLLVerlag": "WOLL Verlag",
    "Wagner": "Wagner Verlag",
    "Wagner Verlag": "Wagner Verlag",
    "Waldgut": "Waldgut Verlag",
    "Waldgut Verlag": "Waldgut Verlag",
    "Wallstein": "Wallstein Verlag",
    "Wallstein Verlag": "Wallstein Verlag",
    "Wallstein-Verl.": "Wallstein Verlag",
    "Wartberg": "Wartberg Verlag",
    "Wartberg Verlag": "Wartberg Verlag",
    "Wartberg-Verl.": "Wartberg Verlag",
    "Wartberg-Verlag GmbH": "Wartberg Verlag",
    "Wartburg Verlag": "Wartburg Verlag",
    "Wartburg-Verl.": "Wartburg Verlag",
    "Wehrhahn": "Wehrhahn Verlag",
    "Wehrhahn Verlag": "Wehrhahn Verlag",
    "Wehrhahn-Verl.": "Wehrhahn Verlag",
    "Weimarer Schiller Presse": "Weimarer Schiller-Presse",
    "Weimarer Schiller-Presse": "Weimarer Schiller-Presse",
    "Weimarer Schillerpresse": "Weimarer Schiller-Presse",
    "Weimarer-Schiller-Presse": "Weimarer Schiller-Presse",
    "Weishaupt": "Weishaupt Verlag",
    "Weishaupt Verlag": "Weishaupt Verlag",
    "Wellhöfer": "Wellhöfer Verlag",
    "Wellhöfer Verlag": "Wellhöfer Verlag",
    "Wendepunkt Verlag": "Wendepunkt Verlag",
    "Wendepunkt-Verl.": "Wendepunkt Verlag",
    "Werd & Weber Verlag": "Werd & Weber Verlag",
    "Werd & Weber Verlag AG": "Werd & Weber Verlag",
    "Westarp BookOnDemand": "Westarp BookOnDemand",
    "Westarp BookOnDemand in der Mediengruppe Westarp": "Westarp BookOnDemand",
    "Westflügel Verlag": "Westflügel Verlag",
    "Westflügel-Verl.": "Westflügel Verlag",
    "Westkreuz-Verl.": "Westkreuz-Verlag",
    "Westkreuz-Verl. Berlin/Bonn": "Westkreuz-Verlag",
    "Wiebers": "Wiebers Verlag",
    "Wiebers Verlag": "Wiebers Verlag",
    "WiedenVerl.": "Wiedenverlag",
    "Wiedenverl.": "Wiedenverlag",
    "Wiedenverlag": "Wiedenverlag",
    "Wiesenburg": "Wiesenburg Verlag",
    "Wiesenburg Verlag": "Wiesenburg Verlag",
    "Wiesenburg-Verl.": "Wiesenburg Verlag",
    "Wieser": "Wieser Verlag",
    "Wieser Verlag": "Wieser Verlag",
    "Wittgenstein Verlag": "Wittgenstein Verlag",
    "Wittgenstein-Verl.": "Wittgenstein Verlag",
    "Wolfgang Hager Verlag": "Wolfgang Hager Verlag",
    "Wolfgang HagerVerlag": "Wolfgang Hager Verlag",
    "Wortfeger": "Wortfeger Media",
    "Wortfeger Media GmbH": "Wortfeger Media",
    "Wortspieler aus der Gemeindebücherei": "Wortspieler",
    "Autorengruppe Wortspieler": "Wortspieler",
    "Wurdack": "Wurdack Verlag",
    "Wurdack Verlag": "Wurdack Verlag",
    "Zwiebelzwerg": "Zwiebelzwerg Verlag",
    "Zwiebelzwerg Verlag": "Zwiebelzwerg Verlag",
    "Zwiebelzwerg-Verl.": "Zwiebelzwerg Verlag",

    "Adakia-Verl.": "adakia Verlag",
    "adakia Verlag": "adakia Verlag",
    "adakia Verlag UG": "adakia Verlag",
    "Amardi-Verl.": "amardi verlag",
    "amardi verlag": "amardi verlag",
    "AWS Literatur": "aws Literatur",
    "Aws-Literatur, Verlag des Kulturvereins Alles wird schön": "aws Literatur",
    "aws Literatur": "aws Literatur",
    "awsLiteratur": "aws Literatur",
    "Axel Dielmann Verlag": "axel dielmann verlag",
    "Dielmann, Axel": "axel dielmann verlag",
    "axel dielmann - verlag": "axel dielmann verlag",
    "brennt GmbH": "brennt! GmbH",
    "brennt! GmbH": "brennt! GmbH",
    "camino": "camino",
    "camino.": "camino",
    "Capybarabooks": "capybarabooks",
    "capybarabooks": "capybarabooks",
    "custos verlag e.K.": "custos verlag",
    "Custos-Verl.": "custos verlag", 
    "DeBehr, Verlag": "DeBehr",
    "Design, SilvForLife": "SilvForLife Design",
    "Deutscher Lyrik Verlag": "deutscher lyrik verlag",
    "deutscher lyrik verlag": "deutscher lyrik verlag",
    "deutscher lyrik verlag (dlv)": "deutscher lyrik verlag",
    "dtv": "dtv Verlagsgesellschaft",
    "dtv Verlagsgesellschaft": "dtv Verlagsgesellschaft",
    "eFeF-Verl.": "eFeF-Verlag",
    "eFeF-Verlag": "eFeF-Verlag",
    "echomedia Buchverlag": "echomedia buchverlag",
    "echomedia buchverlag": "echomedia buchverlag",
    "Elbaol Verlag": "Elbaol Verlag",
    "Elbaol Verlag Hamburg": "Elbaol Verlag",
    "Elbaol-Verl.": "Elbaol Verlag",
    "elbaol Verlag": "Elbaol Verlag",
    "elbaol verlag": "Elbaol Verlag",
    "elbaol verlag hamburg": "Elbaol Verlag",
    "hockebooks": "hockebooks",
    "hockebooks gmbh": "hockebooks",
    "fe-Medienverlags-GmbH": "fe-medienverlag",
    "fe-medienverl.": "fe-medienverlag",
    "felu": "felu Verlag",
    "felu Verlag GbR - Ketti Gnida u. Susanne Krebin": "felu Verlag",
    "fhl Verl. Leipzig UG": "fhl Verlag",
    "fhl-Verl.": "fhl Verlag",
    "Ihleo-Verl.-Büro": "ihleo verlag",
    "ihleo Verlag": "ihleo verlag",
    "ihleo verlag": "ihleo verlag",
    "ihleo verlagsbüro - Dr. Oliver Ihle": "ihleo verlag",
    "Konkursbuch Verlag Claudia Gehrke": "konkursbuch verlag",
    "Konkursbuch-Verl. Gehrke": "konkursbuch verlag",
    "konkursbuch": "konkursbuch verlag",
    "konkursbuch Verlag Claudia Gehrke": "konkursbuch verlag",
    "Marixverl.": "marixverlag",
    "S. Marix Verlag": "marixverlag",
    "marix Verlag": "marixverlag",
    "marix verlag": "marixverlag",
    "marixverlag": "marixverlag",
    "OHNEOHREN": "ohneohren verlag",
    "Ohneohren-Verl.": "ohneohren verlag",
    "Verlag ohneohren": "ohneohren verlag",
    "ohneohren Verlag": "ohneohren verlag",
    "ohneohren verlag": "ohneohren verlag",
    "Onomato-Verl.": "onomato verlag",
    "onomato Verlag": "onomato verlag",
    "onomato verlag": "onomato verlag",
    "p.machinery": "p.machinery",
    "p.machinery Haitel": "p.machinery",
    "p.machinery Haitel***5173734": "p.machinery",
    "p.machinery Michael Haitel": "p.machinery",
    "public book media Verlag": "public book media verlag",
    "public book media verlag": "public book media verlag",
    "Sonderzahl": "sonderzahl",
    "sonderzahl": "sonderzahl",
    "stolze augen books": "stolze augen books",
    "stolzeaugen.books": "stolze augen books",
    "Stolzeaugen.books UG": "stolze augen books",
    "stolzeaugen.books UG": "stolze augen books",
    "synergenVerlag": "synergenVerlag",
    "synergenVerlag - Deutscher Bierkulturverlag": "synergenVerlag",
    "tolino media": "tolino media",
    "via tolini media": "tolino media",
    "via tolino media": "tolino media",
    "via tolino-media": "tolino media",
    "tpk-Regionalverlag": "tpk-Verlag",
    "tpk-Verlag": "tpk-Verlag",
    "tredition": "tredition",
    "tredition GmbH": "tredition",
    "Tredition GmbH &lt": "tredition",
    "Trötsch Verlag": "Trötsch Verlag GmbH & Co. KG",
    "Trötsch Verlag GmbH & Co.KG, Trötsch Verlag GmbH &.": "Trötsch Verlag GmbH & Co. KG",
    "wesText Verlag": "wesText Verlag",
    "wesText-Verl.": "wesText Verlag",
    "Verlag xlibri": "xlibri",
    "xlibri": "xlibri",
    "xlibri.de": "xlibri",
    "zu Klampen": "zu Klampen!",
    "zuKlampen!": "zu Klampen!",
    "Éd.  Saint-Paul": "Éd. Saint-Paul",
    "Éd. Saint-Paul": "Éd. Saint-Paul",
    "Österr. Literaturforum": "Österreichisches Literaturforum",
    "Österreichisches Literaturforum": "Österreichisches Literaturforum",
    "Agiro": "AGIRO Verlag",
    "Arachne-Verl.": "Arachne Verlag",
    "Baeschlin": "Baeschlin Verlag",
    "BasisDruck": "BasisDruck Verlag",
    "Bastei Entertainment": "Bastei Lübbe",
    "Bastei Lübbe Taschenbuch": "Bastei Lübbe",
    "be.bra Verlag": "Be.bra-Verl.",
    "Birgit Otten": "Birgit Otten c/o Papyrus Autoren-Club",
    "Block": "Block-Verlag",
    "Blumenbar": "Blumenbar-Verl.",
    "Boer, K": "Boer Verlag",
    "Bookspot Edition 211": "Bookspot-Verl.",
    "BOOKSun-Limited, Niederlassung Dt.": "BOOKSun-Limited",
    "BördeAutoren e.V., BördeAutoren": "BördeAutoren e.V.",
    "BördeAutoren, Die": "BördeAutoren e.V.",
    "BS-Verl.": "BS Verlag Rostock",
    "BS-Verl. Rostock": "BS Verlag Rostock",
    "BS-Verl.-Rostock": "BS Verlag Rostock",
    "BS-Verl.-Rostock Bruhn": "BS Verlag Rostock",
    "BS-Verlag": "BS Verlag Rostock",
    "BS-Verlag-Rostock": "BS Verlag Rostock",
    "BS-Verlag-Rostock Angelika Bruhn": "BS Verlag Rostock",
    "buch & media": "Buch&Media",
    "Buchwerft-Verl.": "buchwerft-verlag.de",
    "Chiliverlag": "chiliverlag",
    "concepcion Seidel OHG": "Concepcion Seidel OHG",
    "copy-us Verlags GmbH": "Copy-us-Verl.-GmbH",
    "Das erotische Sekretariat": "Das Erotische Sekretariat",
    "Demmler": "DEMMLER VERLAG",
    "der wolf verlag": "Der Wolf Verlag",
    "deutscher lyrik verlag": "Deutscher Lyrik Verlag",
    "Dialog Ed.": "Dialogverlag",
    "dialogverlag": "Dialogverlag",
    "dr. ziethen verlag": "Dr. Ziethen Verlag",
    "Drachenstern Verlag": "DrachenStern Verlag",
    "echomedia buchverlag": "Echomedia-Buchverl.",
    "edition TIAMAT": "Edition TIAMAT",
    "Ed. Tiamat": "Edition TIAMAT",
    "Edition Tintenfaß": "Edition Tintenfass",
    "Ed. Tintenfaß": "Edition Tintenfass",
    "Ed. Tirol": "Edition Tirol",
    "Edition Tirol Martin Reiter": "Edition Tirol",
    "Ed. Töpfl": "Edition Töpfl",
    "edition 8": "Edition 8",
    "Ed. 8": "Edition 8",
    "edition a": "Ed. A",
    "Edition Atelier": "Ed. Atelier",
    "Ed. Contra-Bass": "Edition Contra-Bass",
    "edition exil": "Ed. Exil",
    "Edition Eizenbergerhof": "Ed. Eizenbergerhof",
    "edition eizenbergerhof": "Ed. Eizenbergerhof",
    "edition sonne & mond": "edition sonne und mond",
    "edition q im be.bra verlag": "edition q im bebra verlag",
    "Edition Federleicht": "edition federleicht",
    "Edition FZA": "edition fza",
    "Edition Howeg": "Ed. Howeg",
    "Edition Hübscher": "Ed. Hübscher",
    "Ed. Geheimes Wissen": "Edition Geheimes Wissen",
    "Edition Isele": "Ed. Isele",
    "Isele": "Ed. Isele",
    "Edition Keiper": "edition keiper",
    "Edition Köndgen": "Ed. Köndgen",
    "Edition Lagarto": "Ed. Lagarto",
    "edition lex liszt 12": "Ed. lex liszt 12",
    "edition lex liszt 12 Verein zur Foerderung": "Ed. lex liszt 12",
    "edition lexl iszt12": "Ed. lex liszt 12",
    "Edition Märkische LebensArt": "Ed. Märkische Lebensart",
    "Edition Meerauge": "Ed. Meerauge",
    "Edition Narrenflug": "Ed. Narrenflug",
    "Edition Noack & Block": "Ed. Noack & Block",
    "edition nove": "Ed. Nove",
    "Edition Paashaas Verlag": "Ed. Paashaas Verlag",
    "edition punktuell.": "Ed. Punktuell",
    "Ed. Roesner": "Edition Roesner",
    "Edition Roter Drache": "Ed. Roter Drache",
    "Edition Schwarzdruck": "Ed. Schwarzdr.",
    "Edition Signathur": "Ed. Signathur",
    "Edition Subkultur": "Ed. Subkultur",
    "edition.subkultur.de": "Ed. Subkultur",
    "Editions Saint-Paul": "Éd. Saint-Paul",
    "Edition Tandem": "Ed. Tandem",
    "Edition Temmen": "Ed. Temmen",
    "Edition Wilderermuseum": "Ed. Wilderermuseum",
    "Verlag Wilderermuseum, Verlag": "Ed. Wilderermuseum",
    "Edition Wendepunkt": "Ed. Wendepunkt",
    "Edition Text und Media": "Ed. Text und Media",
    "edition worte & werke": "edition worte&werke",
    "EDITION XXL GmbH": "EDITION XXL",
    "Edition Saarländisches Künstlerhaus": "Ed. Saarländisches Künstlerhaus",
    "Eifeler Literaturverlag": "ELV, Eifeler Literaturverlag",
    "EINBUCH Buch- und Literaturverlag": "Einbuch Buch- und Literaturverlag",
    "ELVEA": "Elvea Verlag",
    "Verlag, Elvea": "Elvea Verlag",
    "Verlag, Kelebek": "Kelebek Verlag",
    "EPLA Verlag": "Epla-Verlag",
    "Fabrik.transit": "fabrik.transit",
    "Fe-Medien": "fe-medienverlag",
    "fhl Verlag": "FHL-Taschenbuch-Verl.",
    "Gerhard Hess Verlag": "Gerhard-Hess-Verlag",
    "GROH": "Groh Verlag",
    "Hess": "Hess Verlag",
    "Hkl-Verl.": "HKL-Verlag",
    "Hochroth": "hochroth",
    "hora Verlag": "Hora-Verl.",
    "Hüsch & Hüsch": "Hüsch & Hüsch GmbH",
    "Iatros Verlag": "IATROS-Verlag und Services GmbH",
    "IATROS Verlag & Services GmbH": "IATROS-Verlag und Services GmbH",
    "Ibidem-Verl.": "ibidem-Verlag",
    "ikon VerlagsGesmbH": "Ikon-Verl.-GesmbH",
    "INNOCENTIA VERLAG Rainer Hüls e.K.": "Innocentia Verlag",
    "INNSALZ": "Innsalz",
    "innsalz": "Innsalz",
    "Jmb-Verl.": "jmb-Verlag",
    "J. S. Klotz Verlagshaus": "J.S. Klotz Verlagshaus",
    "Jung Wien '14 &lt": "Jung Wien '14",
    "JVA Hannover, Inhaftierte der": "JVA Hannover", 
    "JVA Hannover, Projekt CaRIBu in der": "JVA Hannover", 
    "Kawohl": "Kawohl Verlag",
    "Kampa": "Kampa Verlag",
    "KILLROY media": "Killroy Media",
    "KITAB": "Kitab",
    "Klever Verlag": "Klever-Verl.",
    "KommPlot, AutorInnengruppe": "KommPlot, Autorinnengruppe",
    "KSB-Media": "KSB-Media GmbH",
    "KUUUK, Verlag mit 3 U": "KUUUK-Verl.",
    "Laurin": "laurin",
    "Lebema.ch": "lebema.ch",
    "lectorbooks": "lectorbooks, ein Imprint der Torat GmbH",
    "Lehmanns Media": "lehmanns media",
    "LIT": "Lit",
    "Mandelbaum Verlag": "mandelbaum verlag",
    "Mare": "mare",
    "maringo": "Maringo Verlag",
    "MediaKern": "mediaKern",
    "Modo": "modo",
    "Morgenroth Media": "morgenroth media",
    "Moses": "moses Verlag",
    "Muc Verlag": "muc Verlag",
    "Münstermann": "münstermann",
    "Müry Salzmann": "müry salzmann",
    "myMorawa": "Mymorawa von Morawa Lesezirkel GmbH",
    "Nicolai": "nicolai",
    "Noel-Verl.": "NOEL-Verlag",
    "Odenwaldkreises, Kreisausschuss des": "Odenwaldkreis, Kreisausschuss",
    "Offizin Verlag": "OFFIZIN Zürich Verlag",
    "Pandion": "Pandion Verlag",
    "Patmos": "Patmos Verlag",
    "piepmatz Verlag": "Piepmatz-Verl.",
    "pkp Verlag, Pierre Kynast": "pkp Verlag",
    "Pkp-Verl.": "pkp Verlag",
    "Plattform": "plattform",
    "PLV": "plV-Media",
    "Poss": "POSS Verlag",
    "Projekt Verlag": "projektverlag",
    "Public book media Verlag": "public book media verlag",
    "PUPUT BOOKS": "Puput Books",
    "Quartus-Verl.": "quartus-Verlag",
    "RADIUS": "Radius Verlag",
    "Ratio Books": "ratio-books",
    "Regia": "REGIA Verlag",
    "Riva": "riva",
    "riva Verlag": "riva",
    "Romantruhe": "Romantruhe - Buchversand",
    "Rosenheimer": "rosenheimer",
    "Salon LiteraturVerlag": "Salon-LiteraturVerl.",
    "SALONLiteraturVERLAG": "Salon-LiteraturVerl.",
    "SATYR Verlag": "Satyr-Verl.",
    "Seifert": "Seifert Verlag",
    "Sell, Heimat-Verl.": "S. Sell Heimat-Verlag",
    "servitus Verlag": "Servitus-Verl.",
    "Skript-Verlag": "skript-Verlag",
    "sonderpunkt Verlag": "Sonderpunkt-Verl.",
    "Sujet Verlag": "sujet verlag",
    "Telegonos-publishing": "telegonos-publishing",
    "TextKontor": "Textkontor",
    "Trafo": "trafo",
    "Tunkan": "tunkan",
    "tvd": "tvd-Verl.",
    "Unartproduktion": "unartproduktion",
    "Und-Verl.": "UND-Verlag",
    "van Eck": "van Eck Verlag",
    "VS Bayern - Verband Deutscher Schriftstellerinnen und Schriftsteller in Bayern / Regionalgruppe Ostbayern": "Verband Deutscher Schriftsteller / Regionalgruppe Ostbayern",
    "Verl. Die Werkstatt": "Verlag Die Werkstatt",
    "Verl. die Werkstatt": "Verlag Die Werkstatt",
    "Verlag make a book": "Verl. Make a Book",
    "Verl. Make a book": "Verl. Make a Book",
    "Verl. Rotabene Medienhaus": "Verl. Rotabene-Medienhaus",
    "Verlag Akademie der Abenteuer": "Verlag Akademie-der-Abenteuer",
    "Verlag am Park": "verlag am park",
    "Verl. am Park": "verlag am park",
    "Verlag der Neun Reiche": "Verlag der neun Reiche",
    "Verlag Traugott Bautz": "Verlag Traugott Bautz GmbH",
    "Verlagsallianz": "verlagsallianz",
    "Verlag 3.0": "Verl. 3.0",
    "Verl. am Eschbach": "Verlag am Eschbach",
    "Verl. am Schloss": "Verlag am Schloss",
    "Verl. Atelier im Bauernhaus": "Verlag Atelier im Bauernhaus",
    "Verl. Begegnungen": "Verlag Begegnungen",
    "Verl. Blaues Schloss": "Verlag Blaues Schloss",
    "Verl. BS": "Verlag BS",
    "Verl. CCU": "Verlag CCU",
    "Verl. Der Apfel": "Verlag Der Apfel",
    "Verl. der Schatten": "Verlag der Schatten",
    "Verl. der Villa Fledermaus": "Verlag der Villa Fledermaus",
    "Verl. des Biographiezentrums": "Verlag des Biographiezentrums",
    "Verl. Die Blaue Eule": "Verlag Die Blaue Eule",
    "Verl. Die Brotsuppe": "verlag die brotsuppe",
    "verlag der brotsuppe": "verlag die brotsuppe",
    "Verlag Edition AV": "Verl. Ed. AV",
    "Verl. Expeditionen": "Verlag Expeditionen",
    "Verl. Fischueberalles.ch": "Verlag fischueberalles.ch",
    "Verlag für Kultur und Wissenschaft Dr. Thomas Schirrmacher": "Verl. für Kultur und Wiss.",
    "Verlag Herder": "Herder",
    "Verl. Herder": "Herder",
    "Verl. im Inst. Drachenhaus": "Verlag im Institut Drachenhaus",
    "Verl. Rote Zahlen": "Verlag Rote Zahlen",
    "Verl.-Allianz": "verlagsallianz",
    "VERRAI-VERLAG": "Verrai-Verlag",
    "VHV, Literatur & Kultur Verlag": "vHVerlag",
    "Vorarlberg> / Kultur Referat": "Kultur Referat Vorarlberg",
    "Weissbooks.w": "weissbooks.w",
    "Wißner": "Wißner-Verlag",
    "Wolf Präsente": "Wolf Präsente GmbH",
    "WOLL Verlag": "Woll Verlag",
    "WOLL-Verlag Hermann-J. Hoffe": "Woll Verlag",
    "Xenomoi": "xenomoi",
    "ZWIEBOOK": "zwiebook",
    "Zwiebook": "zwiebook",
    "Zytglogge": "Zytglogge Verlag",
}

# Funktion zur Konsolidierung der Namen anhand des Mappings
def konsolidiere_und_speichere(file_path, columns_to_map, mapping, save_dir):
    df = pd.read_csv(file_path, sep=';', encoding='utf-8')

    base = os.path.basename(file_path)
    file_root, ext = os.path.splitext(base)

    for column_name in columns_to_map:
        if column_name in df.columns:
            # Ausgabe der ursprünglichen Namen zur Debugging-Zwecken
            print(f"Datei: {base}, Spalte: {column_name}")
            print("Ursprüngliche Namen: ", df[column_name].dropna().unique()[:5])

            konsolidierte_names = []

            for eintrag in df[column_name]:
                if pd.notna(eintrag):
                    names = [name.strip() for name in eintrag.split(';') if name.strip()]
                    konsolidierte_names.append(';'.join(mapping.get(name, name) for name in names))
                else:
                    konsolidierte_names.append(eintrag)
            
            df[column_name] = konsolidierte_names

            # Ausgabe der konsolidierten Namen zur Debugging-Zwecken
            print("Konsolidierte Namen: ", df[column_name].dropna().unique()[:5])

        # Speichere immer nach der Bearbeitung der letzten Spalte
        neuer_dateiname = os.path.join(save_dir, f"{file_root}_kk{ext}")
        df.to_csv(neuer_dateiname, index=False, encoding='utf-8', sep=';')
        print(f"✅ Datei {neuer_dateiname} erfolgreich konsolidiert und gespeichert.")

# Definiere den Speicherort für die konsolidierten Dateien
save_directory = r'C:\Pfad\zu\neuem\Verzeichnis\Konsolidierte'

# Sorge dafür, dass das Verzeichnis existiert
os.makedirs(save_directory, exist_ok=True)

# Dateipfade und Spalten zur Konsolidierung (gegebene CSV könnte mehrere Spalten enthalten)
dateien = [
    (r'C:\Pfad\zur\2008-2023_Buch.csv', ['buchverlag', 'verfasser', 'herausgeber', 'uebersetzer', 'keinerolle', 'sonstige']),
    (r'C:\Pfad\zur\Literaturwettbewerb.csv', ['veranstalter']),
    (r'C:\Pfad\zur\Literaturwettbewerb_Ausgabe.csv', ['gewinner', 'jury']),
    (r'C:\Pfad\zur\Literaturzeitschrift.csv', ['zeitschriftverlegtvon', 'creator']),
    (r'C:\Pfad\zur\Literaturzeitschrift_Ausgabe.csv', ['verfasser_ke_beiträge']),
    (r'C:\Pfad\zur\Plattform_Internet.csv', ['unternehmen'])
]


# Konsolidiere alle Dateien
for datei, spalten in dateien:
    konsolidiere_und_speichere(datei, spalten, koerperschaft_mapping, save_directory)
