==> skripte_datenkorpus/6_Konsolidierung_Person.py <==
import os
import pandas as pd

# Mapping von inkonsistenten zu konsistenten Personennamen
personenname_mapping = {
"#wirschreibenfürahrweiler - Band für Ahrweiler Kindergarten SWR Herzenssache, Autorengemeinschaft": "wirschreibenfürahrweiler, Autorengemeinschaft",
"#wirschreibenfürahrweiler - Band für Aktion hilft Deutschland, Autorengemeinschaft": "wirschreibenfürahrweiler, Autorengemeinschaft",
"(Hg.), Franz-Joseph Huainigg": "Huainigg, Franz-Joseph",
"(Hg.), Johann Stockinger": "Stockinger, Johann",
"(Hrsg.), Treffpunkt Schreiben": "Treffpunkt Schreiben",
"(Hsg.), Ella Stein &. Tom U. Behrens": "Stein, Ella; Behrens, Tom U.",
"«Henry» Lüscher, Heiner": "Lüscher, Heiner", 
"«Joggi» Jost, Peter": "Jost, Peter",
"Allen, Lucius": "Allan, Lucius",
"Arnim, Ludwig Achim von": "Arnim, Achim von",
"Arnolds, Dr., Günter": "Arnolds, Günter",
"Athanassiou, Dimitros": "Athanassiou, Dimitrios",
"Block, Detlev Geistiger Schöpfer": "Block, Detlev",
"Bodroži?, Marica": "Bodrožić, Marica", 
"Boose und andere, Ruth und andere": "Boose, Ruth",
"C, Eloise ..": "C., Eloise",
"Cernhuby, Stefan": "Cernohuby, Stefan",
"Cosette Cosette": "Cosette", 
"Dijzmedjian, Marc": "Djizmedjian, Marc",
"Döring, Volkmar O.": "Döring, Volkmar", 
"Falak?, Ma?m?d": "Falaki, Mahmood", 
"Hagen, Dr. Barbara": "Hagen, Barbara",
"Hoffmann, Jannika": "Hoffmann, Janika",
"Hoffmann, Marius": "Hoffmann, Mariusz", 
"Holms, Ernerstine": "Holms, Ernestine", 
"Joosten, Ullrich": "Joosten, Ulrich",
"Kafka, František": "Kafka, Franz", 
"Kleine Stevermüer, Kleine": "Kleine Stevermüer, Ursula",
"Lang, Sybille": "Lang, Sibylle",
"Aboreas, Rüdiger N.": "Aboreas, Rüdiger",
"Aellig, Pierre E.": "Aellig, Pierre Emile",
"Alavie, Aye": "Alavie, Ayeda",
"Albrecht, Susanne Maria Ulrike": "Albrecht, Susanne Ulrike Maria",
"Annecke, Martina": "Anneke, Martina",
"Arbaoui, Cornelia": "Arbaoui, Cornelia Erika",
"Auth, Gabi": "Auth, Gabriele",
"Balaka, Bettina": "Balàka, Bettina",
"Barth, Heinz G.": "Barth, Heinz-Georg",
"Baumann, Evgenija": "Baumann, Evgenija A.",
"Brüstle, Ruben A.": "Brüstle, Ruben",
"Brüstle, Rubens": "Brüstle, Ruben",
"Bulenda, Doris E.M.": "Bulenda, Doris E. M.",
"Burroughs, Edgar Ric": "Burroughs, Edgar Rice",
"Bürvenich, Thomas": "Bürvenich, Thomas J.",
"Cremer, Marks": "Cremer, Markus",
"De La Marre, Wentila": "De la Marre, Wentila",
"Dell’Aere, Annette": "Dell'Aere, Annette",
"Deyval, Ben A.": "Deyval, Benjamin Alexander",
"do Rego Silva, Tita": "do Rêgo Silva, Tita",
"Dobrovolny-Mühlenbach, Petra": "Dobrovolny-Mühlenbach, Petra M.",
"Drewing, Ingrid": "Drewing, Ingrid Herta",
"Droullier, Jacqueline": "Droullier, Jacqueline V.",
"Du Pre, Corinne": "Du Pré, Corinne",
"Edelbrock, MIchael": "Edelbrock, Michael",
"Elgass, Peter": "Elgaß, Peter",
"Engel, Rénée": "Engel, Renée",
"Esser, Angela": "Eßer, Angela",
"Fichtl, Betty": "Fichtl, Betti",
"Fillbrandt, Hildegart": "Fillbrandt, Hildegard",
"Finger, Berhard": "Finger, Bernhard",
"Fitzgerald, Francis Scott Key": "Fitzgerald, F. Scott",
"Fleischer, Ludwig": "Fleischer, Ludwig Roman",
"Fritsch, Valerie Katrin G.": "Fritsch, Valerie",
"Gabon, Ilvy Marie": "Gabon, Ilvy-Marie",
"Gahlen, Sarah": "Gahlen, Sarah Fiona",
"Gillespie, Jaqueline": "Gillespie, Jacqueline",
"Goldenbaum, Gianna S.": "Goldenbaum, Gianna Suzann",
"GÖLZENLEUCHTER, HORST DIETER": "Gölzenleuchter, Horst-Dieter",
"Grausgruber, Gabi": "Grausgruber, Gabriele",
"Grobe, Dietrich W.": "Grobe, Dietrich Wilhelm",
"Gross, Angelika": "Groß, Angelika",
"Grossmann, Ulf": "Großmann, Ulf",
"Handl, Heimo L.": "Handl, Haimo L.",
"Hary, W. A.": "Hary, Wilfried A.",
"Haslehner, Elfriede": "Haslehner-Götz, Elfriede",
"Haußmann (Hg), Renate": "Haußmann, Renate",
"Heintel, Lars": "Heintel, Lars O.",
"Heizmann, Jeanpiere": "Heizmann, Jeanpierre",
"Heller, Dorina": "Heller, Dorina M.",
"Helmers, Astrid": "Helmers, Astrid M.",
"Hetterling, Hans Jürgen": "Hetterling, Hans-Jürgen",
"Hidringer, Christine": "Hidringer, Christine A.",
"Hnidek Leopold, Hnidek": "Hnidek, Leopold",
"Hoffmann, E. T. A.": "Hoffmann, E.T.A.",
"Höfler, Maria Magda": "Höfler, Maria Magdalena",
"Höpflinger, Anna": "Höpflinger, Anna-Katharina",
"Ihnken, Barbara": "Ihnken, Iris Barbara",
"Innocenti, Alexa": "Innocenti, Alexandra", 
"Juyub, Q.A.": "Juyub, Qayid Aljaysh",
"Kämpf, Mändy": "Kämpf, Mandy",
"Kappel, Charly": "Kappel, Charlie",
"KArg, Thomas": "Karg, Thomas",
"katzur, Christa": "Katzur, Christa",
"Kawka-Wegmann, Annie": "Kawka-Wegmann, Annie Katharina",
"Kemps des Escalante, Caroline": "Kemps de Escalante, Caroline",
"Kensche, ElviEra": "Kensche, Elviera", 
"Kloth, Franziska": "Kloth, Franziska-Gabriele",
"Kogler, Iris": "Kogler, Iris Antonia",
"Koschorrek, Anita": "Koschorrek-Müller, Anita",
"Lendle,  Jo": "Lendle, Jo",
"Loschütz, Gerd": "Loschütz, Gert",
"Poe, E.T.": "Poe, Edgar Allan", 
"Polt, Gerhardt": "Polt, Gerhard",
"Puškin, Aleksandr Sergeevi?": "Puschkin, Alexander",
"Rammsted, Tilman": "Rammstedt, Tilman",
"Rilke, Reiner M.": "Rilke, Rainer Maria",
"Kurdi, Hartmut El": "El Kurdi, Hartmut",
"Kurdi, Hartmut el-": "El Kurdi, Hartmut",
"Kurfürst-Meins, Gisela": "Kürfürst-Meins, Gisela",
"Ley, Roy F.": "Ley, Roy Francis",
"Lipp, Wilhelm Maria": "Lipp, Wilhelm-Maria",
"Lips, Yvi": "Lips, Yvy",
"Lovecraft, H. P.": "Lovecraft, Howard Phillips",
"Lovecraft, H.P.": "Lovecraft, Howard Phillips",
"Ludwig, M. W.": "Ludwig, M.W.",
"MaaßEn, Volker": "Maaßen, Volker",
"Mandt, Silvia": "Mandt, Sylvia",
"Marian, Cornelie": "Marian, Cornelíe",
"Marquis de Sade, Donatien": "Marquis de Sade, Donatien-Alphonse-François",
"Meinicke, Claudia G.": "Meinicke, Claudia Gabriele",
"Meiser, Oliver": "Meiser, Olivér",
"Minden, Inka L.": "Minden, Inka Loreen",
"Montemurri, Jaqueline": "Montemurri, Jacqueline",
"Moser, Annemarie": "Moser, Annemarie E.",
"Neder-Goltz, Anna": "Neder von der Goltz, Anna",
"Neder- von der Goltz, Anna": "Neder von der Goltz, Anna",
"Neder-von der Goltz, Anna": "Neder von der Goltz, Anna",
"Neudert, Phillip Emanuel": "Neudert, Philipp Emanuel",
"Nussbaum, Hannelore": "Nußbaum, Hannelore",
"Olsson, Sven j.": "Olsson, Sven J.",
"Orgel, T. S.": "Orgel, T.S.",
"Persson, T. B.": "Persson, Terry B.",
"Peters, Andreas": "Peters, Andreas Andrej",
"Peters, Joachim": "Peters, Joachim H.",
"Plepelits, Karl": "Plepeltis, Karl",
"Pollok, Tim": "Pollock, Tim",
"Prüfer, Jan Christoph": "Prüfer, Jan-Christoph",
"Rabaza, Sonja S.": "Rabaza, Sonja Sophia",
"Rank, Klaudia": "Rank, Klaudia Gräfin von",
"Rauch, Marco": "Rauch, Marco A.",
"Rönisch, Marie-Luise": "Rönisch, Marie-Luis",
"Roth, HC": "Roth, H. C.",
"Ryzmbowski, Susanne": "Rzymbowski, Susanne",
"Salmon, K. K.": "Salmon, Kim Katharina",
"Schech, K. E.": "Schech, Kaspar Eduard",
"Skora, Andre": "Skora, André",
"Spieker, N. S. H.": "Spieker, N.S.H.",
"Staniši?, Saša": "Stanišić, Saša",
"Stavari?, Michael": "Stavaric, Michael",
"Steinmetz, M. H.": "Steinmetz, M.H.",
"Sültz, Uwe H.": "Sültz, Uwe Heinz",
"Thomas, P. C.": "Thomas, P.C.",
"Tolstoi, Lew": "Tolstoi, Leo",
"Verollet, Misha": "Vérollet, Mischa-Sarim",
"Vogt, Judith": "Vogt, Judith C.",
"Vojnov, Dimit?r": "Vojnov, Dimitri",
"Völker, Martin": "Völker, Martin A.",
"Aster, Christian von": "von Aster, Christian",
"Bülow, G. G. von": "von Bülow, G.G.",
"Düffel, John von": "von Düffel, John",
"Ebner-Eschenbach, Marie von": "von Ebner-Eschenbach, Marie",
"Hofmannsthal, Hugo von": "von Hofmannsthal, Hugo",
"Horváth, Ödön von": "von Horváth, Ödön",
"Kalm, Eva von": "von Kalm, Eva",
"Kleist, Eva von": "von Kleist, Eva",
"Kleist, Heinrich von": "von Kleist, Heinrich",
"Lethe, Eris von": "von Lethe, Eris",
"Schlegell, Mark von": "von Schlegell, Mark",
"Steinaecker, Thomas von": "von Steinaecker, Thomas",
"Tingen, Ötti von": "von Tingen, Ötti",
"Wedelstedt, Kuno Georg von": "von Wedelstedt, Kuno Georg",
"Vox, Greta L.": "Vox, Greta Luisa",
"WARDENGA, R.G.": "Wardenga, Renate Gertrud",
"Wardenga, R.G.": "Wardenga, Renate Gertrud",
"Wegler, Ronni": "Wegler, Ronny",
"Wells, Benedikt": "Wells, Benedict",
"Wells, H. G.": "Wells, H.G.",
"Wiesler, Andre": "Wiesler, André",
"Wilberz, Frank": "Wilberz, Frank J.",
"Wildenhain, Luise$cvon": "von Wildenhain, Luise", 
"Wildenhain, Luise von": "von Wildenhain, Luise", 
"Witte-Pflanz, Corina": "Witte-Pflanz, Corinna",
"Wozniak, Waltraud": "Wozniak, Waltraut",
"Wurst-Hanssen, Johanne M.": "Wurst-Hanssen, Johanne Magdalene",
"Zawrel, Renate": "Zawrel, Reni",
"Zegelman, Anne": "Zegelmann, Anne",
"Zelezny, Sabrina": "Železný, Sabrina",

"A. Amsler": "Amsler, André",
"A. Büttner": "Büttner, Anne",
"A. F. W. Sommer": "Sommer, Anton F. W.", 
"A. Gasser": "Gasser, Alex", 
"A. Hovestadt": "Hovestadt, Alfred", 
"A. Langkath": "Langkath, Angela", 
"A. Thorn": "Thorn, Almud",
"A. Tschirner": "Tschirner, Andrea",
"A. Wellbrock": "Wellbrock, Achim",
"A. Westermann": "Westermann, Alfons", 
"A. Zmyj": "Zmyj, Annette",
"Alfred Starch": "Starch, Alfred",
"Andrea Herrmann": "Herrmann, Andrea",
"Andrea Yildiz": "Yildiz, Andrea",
"Andreas Dölling": "Dölling, Andreas",
"Andreas Herbert Meier": "Meier, Andreas Herbert",
"Angela Suter": "Suter, Angela",
"Angelika Horn": "Horn, Angelika", 
"Angelika Bohn, c/o Papyrus Autoren-Club": "Bohn, Angelika",
"Anja M. Kast": "Kast, Anja M.",
"Anke Redhead": "Redhead, Anke",
"Annegret Achner": "Achner, Annegret",
"Annette Kipnowski (Künstlerin) und Marion Raschke (Germanistin)": "Kipnowski, Annette; Raschke, Marion",
"Anselm Eder": "Eder, Anselm", 
"Antje Meeser": "Meeser, Antje",
"Anton Launer": "Launer, Anton",
"Antonia Löschner": "Löschner, Antonia",
"Antonia Wimmer": "Wimmer, Antonia",
"Armin Andrä": "Andrä, Armin",
"Arne Stoll": "Stoll, Arne", 
"B. Dethloff": "Dethloff, Bettina",
"B. E. Bohmann": "Bohmann, Birgit Evi",
"B. Krauß": "Krauß, Barbara",
"B. Leuz": "Schmid-Leuz, Barbara",
"B. Luipold": "Luipold, Bernhard",
"B. Pallina/Natalie Milke c/o Autorenglück": "Pallina, B.; Milke, Natalie",
"B. Schwarz": "Schwarz, Bernhard",
"B. Stählin-Tschanz": "Stählin-Tschanz, Betli",
"Barbara Seelk": "Seelk, Barbara",
"Bauer, Hilde": "Bauer, Hilde",
"Bernd Apel": "Apel, Bernd",
"Bernhard Hoppe": "Hoppe, Bernhard",
"Brigitte Hellebrand": "Hellebrand, Brigitte",
"Brigitte Regitz": "Regitz, Brigitte",
"Brigitte van Hattem": "Brigitte van Hattem",
"Brita Linde": "Linde, Brita",
"Britta Bendixen": "Bendixen, Britta",
"Bruno Moebius": "Moebius, Bruno",
"Bungert, Klauspeter": "Bungert, Klauspeter",
"C. A. Vinandy": "Vinandy, Charel Aristide",
"C. Anastassiou c/o Atelier am Alsterfleet": "Anastassiou, C.",
"C. Gansen-Hainze": "Gansen-Hainze, Christine",
"C. Greene, c/o autorenglück.de": "Greene, Carolina",
"C. Schartner": "Schartner, Christian",
"C. Wachsmann": "Wachsmann, Christiane",
"C.A. Raaven, c/o Papyrus Autoren Club R.O.M. Logicware GmbH": "Raaven, C. A.",
"Carmen Mayer": "Mayer, Carmen",
"Carmen Silvia Gratias, Baroness von Dornberg": "Gratias, Carmen Silvia",
"Carola Heine": "Heine, Carola",
"Carsten Nichte": "Nichte, Carsten",
"CH. SCHROER": "Schroer, Christopher",
"Christa Maria Buß": "Maria, Christa",
"Christa Schloßmacher": "Schloßmacher, Christa",
"Christel Gilberg": "Gilberg, Christel",
"Christian Grimm": "Grimm, Christian",
"Christian Piesch": "Piesch, Christian",
"Christian Seirer": "Seirer, Christian",
"Christiane Kromp": "Kromp, Christiane",
"Christine Schwarzmüller": "Schwarzmüller, Christine",
"Claudia Berwert, diasign.ch": "Berwert, Claudia",
"Claudia Cliff": "Cliff, Claudia",
"Claudia Freyer": "Freyer, Claudia",
"Claudia Otte": "Otte, Claudia",
"Claudia Wendt": "Wendt, Claudia",
"Claus Beese": "Beese, Claus",
"Cleo Rivers, c/o Gwendolyn Wynter Autorenimpressum": "Rivers, Cleo",
"D. Hartung": "Hartung, D.",
"D. Rauskolb": "Rauskolb, Doris",
"D. von Reeken": "Reeken, Dieter von",
"Dana Müller": "Müller, Dana",
"Daniela Vogel": "Vogel, Daniela",
"Dennerlein, Irmgard": "Dennerlein, Irmgard",
"Diana Brandstetter": "Brandstetter, Diana",
"Dieter Höntsch": "Höntsch, Dieter",
"Dieter von Reeken": "Reeken, Dieter von",
"Doro Koch": "Koch, Doro",
"Dr. Elisabeth Kukral": "Kukral, Elisabeth",
"Dr. Hubert Treml": "Treml, Hubert",
"Dr. Maria Zaffarana": "Zaffarana, Maria",
"Dr. Thorsten Menke, terra autistica": "Menke, Thorsten",
"E. Adler": "Adler, Eckhard",
"E. Blaich": "Blaich, E.",
"E. Frieko": "Frieko, Eva",
"E. Grün": "Grün, Erika",
"E. Schuster": "Schuster, Ewald",
"E. Truninger": "Truninger, Edith",
"Elisabeth Decesso": "Decesso, Elisabeth",
"Elisabeth Rüdinger": "Rüdinger, Elisabeth",
"Elke Krys": "Krys, Elke",
"Elke und Hans-Hermann Beier": "Beier, Elke; Beier, Hans-Hermann",
"Elmar S. Hruby": "Hruby, Elmar S.",
"Eugen Füner": "Füner, Eugen",
"Eugen Grinschuk": "Grinschuk, Eugen",
"Eva Engel": "Engel, Eva",
"Eva Lirot": "Lirot, Eva",
"Eva Maria Klima": "Klima, Eva Maria",
"Ewald Dworak": "Dworak, Ewald",
"F. Irmscher": "Irmscher, Frank",
"F. Sargus": "Sargus, F.",
"F. Westermann": "Westermann, Felix",
"Frank Baranowski": "Baranowski, Frank",
"Frank-Thomas Mitschke": "Mitschke, Frank-Thomas",
"Fritz Ucik": "Ucik, Fritz",
"G. Grob": "Grob, Gisela",
"G. Perkams": "Perkams, Günter",
"G. Pils": "Pils, Georg Hugo Donatus",
"G.-D. Witt": "Witt, Gerd-Dieter",
"Gabriela Hofmann": "Hofmann, Gabriela",
"Gabriele Gail": "Gail, Gabriele",
"Gabriele Kaltofen": "Kaltofen, Gabriele",
"Gerhard Oppel": "Oppel, Gerhard",
"Gerhard Staflinger": "Staflinger, Gerhard",
"Gerhard W.H. Köhn": "Höhn, Gerhard W. H.",
"Gisela Rieger": "Rieger, Gisela",
"Gudrun Schultheiss": "Schultheiss, Gudrun",
"Guido Bruggmann": "Bruggmann, Guido",
"Gunny Catell": "Catell, Gunny",
"Günther Paffrath": "Paffrath, Günther",
"H. Deinhammer": "Deinhammer, Helmut",
"H. Fleischle": "Fleischle, Hans",
"H. Gaicher": "Gaicher, Horst",
"H. Ganz": "Ganz, Horst",
"H. Glatte": "Glatte, Harald",
"H. Haßlauer": "Haßlauer, Hans",
"H. Heinze": "Heinze, Horst",
"H. Helnwein": "Helnwein, Helga",
"Helga Helnwein": "Helnwein, Helga",
"H. Markert": "Markert, Hans",
"H. Pooschke": "Pooschke, Hilde",
"H. Schaper": "Schaper, Henning",
"H. Steinborn": "Steinborn, Heinz",
"H. W. Winkelmann": "Winkelmann, Hartmut W.",
"H.-P. Lorang": "Lorang, Hans-Peter",
"H.-U. Meisel": "Meisel, Ulrich",
"Hajo Blum": "Blum, Hajo",
"Hannah Winkler": "Winkler, Hannah",
"Hans-Peter Bauer": "Bauer, Hans-Peter",
"Harald Brettner": "Brettner, Harald",
"Harald Grunder": "Grunder, Harald",
"Harald Wiesmann": "Wiesmann, Harald",
"Heinrich Thein": "Thein, Heinrich",
"Heinz-Hermann Draab": "Draab, Hermann",
"Heinz-Josef Scherer": "Scherer, Heinz-Josef",
"Helen Hartnagel": "Hartnagel, Helen",
"Helga Demuth": "Demuth, Helga",
"Helmut P. Fleischhauer": "Fleischhauer, Helmut P.",
"Henselowsky Boschmann": "Boschmann, Henselowsky",
"Herbert Schnaibel": "Schnaibel, Herbert",
"Herbert Wetterauer": "Wetterauer, Herbert",
"Hermann Scheffler": "Scheffler, Hermann",
"Hiller, Jakob": "Hiller, Jakob",
"Horst Hartmann": "Hartmann, Horst",
"I. Rössl": "Rössl, Ingrid",
"I. Welzig": "Welzig, Inge",
"Ingo Stauch": "Stauch, Ingo",
"Ingrid Heim": "Heim, Ingrid",
"J. Feuerstein": "Feuerstein, Jenny",
"J. Gottanka": "Gottanka, Josef",
"J. Heinze": "Heinze, Joachim",
"J. Kartheiser": "Kartheiser, Josiane",
"J. Reichelt": "Reichelt, Joachim",
"J. Richtzenhain": "Richtzenhain, Jürgen",
"J. Walther": "Walther, Jan",
"Jes Schön c/o skriptspektor e.U.": "Schön, Jes",
"Jessica Graves": "Graves, Jessica",
"Joa Bosch": "Bosch, Joa",
"Joachim Güntzel": "Güntzel, Joachim",
"Johanna Gerhard & Sarah M. Kempen": "Gerhard, Johanna; Kempen, Sarah M.",
"Jörg Weese": "Weese, Jörg",
"Josef Mugler": "Mugler, Josef",
"Josef Schmaus": "Schmaus, Josef",
"Josef Wieser": "Wieser, Josef",
"Josephine Laloi": "Laloi, Josephine",
"Julia Fischer": "Fischer, Julia",
"Juliane Pröll": "Pröll, Juliane",
"Juliane Reister": "Reister, Juliane",
"Jürgen Berus": "Berus, Jürgen",
"Jürgen Hanik": "Hanik, Jürgen",
"Jürgen Reichardt-Kron": "Kron, J. R.",
"Jürgen Ruszkowski": "Ruszkowski, Jürgen",
"K. E. Schmidt": "Schmidt, Kristina",
"K. Frings": "Frings, Karl",
"Kay Ganahl": "Ganahl, Kay",
"K. Ganahl": "Ganahl, Kay",
"K. Greve": "Greve, Kurt",
"K. Hawkes": "Hawkes, Karin",
"K. Kluge": "Kluge, Karlheinz",
"K. Pfaffeneder": "Pfaffeneder, Klaus",
"K. Sölkner": "Sölkner, Kurt",
"K. W. Herterich": "Herterich, Klaus W.",
"K.-G. Hirschmann": "Hirschmann, Karl-Gustav",
"Kalbhenn, Marlies": "Kalbhenn, Marlies",
"Karin Hackbart": "Hackbart, Karin",
"Karin Peter": "Peter, Karin",
"Karl Aßmann": "Aßmann, Karl",
"Karl Müller": "Müller, Karl",
"Katharina Rambeaud": "Rambeaud, Katharina",
"Kathrin Maier": "Maier, Kathrin",
"Ken Zelenka": "Zelenka, Ken",
"Klaudia Zotzmann-Koch c/o Bianca Kronsteiner, impressumservice.net": "Zotzmann, Klaudia",
"Klauspeter Bungert": "Bungert, Klauspeter",
"Kristin Gruber": "Gruber, Kristin",
"Kristina Lieschke": "Lieschke, Kristina",
"Kurt F. Neubert": "Neubert, Kurt F.",
"Kurt Haberstich": "Haberstich, Kurt",
"Lars Olsson, c/o Block Services": "Olsson, Lars",
"Laura Kier": "Kier, Laura",
"Laura-Marie Andresen": "Andresen, Laura-Marie",
"Lavinia Dierssen": "Dierssen, Lavinia",
"Leon Gutsche": "Gutsche, Leon",
"Lisa Hölzle": "Hölzle, Lisa",
"Lorenz Göddemeyer": "Göddemeyer, Lorenz",
"Lothar Rosenberg": "Rosenberg, Lothar",
"Lukas Böhl": "Böhl, Lukas",
"Luzie Pan": "Pan, Luzie",
"M. Belz": "Belz, Manfred",
"M. Berger": "Berger, M.",
"M. Cane c/o Werneburg Internet Marketing und Publikations-Service": "Cane, M.",
"Martin Frenk": "Frenk, Martin",
"M. Frenk": "Frenk, Martin",
"M. Kraus": "Kraus, Margit",
"M. Lengfelder": "Lengfelder, Martin",
"M. Mauthe": "Mauthe, Margret",
"M. Ritter": "Ritter, Marion",
"M. Schreckenberg, K. Völlmicke und M. @Lamprecht": "Schreckenberg, Monika; Völlmicke, Kristin; Lamprecht, Michael",
"M. Winterhalter": "Winterhalter, Martin",
"Maar, Klaus Oliver": "Maar, Oliver",
"Madame Saskia": "Saskia, Madame",
"Madeleine Puljic": "Puljic, Madeleine",
"Maik Lindner": "Lindner, Maik",
"Mandy Passehl": "Passehl, Mandy",
"Manuela Thiel": "Thiel, Manuela",
"Marcel Kuiper": "Kuiper, Marcel",
"Marcel Porta": "Porta, Marcel",
"Marco Rochholz": "Rochholz, Marco",
"Maria Föderl": "Föderl, Maria",
"Maria Meyer": "Meyer, Maria",
"Marina Ocean": "Ocean, Marina",
"Marina Schuster": "Schuster, Marina",
"Marion Kummerow": "Kummerow, Marion",
"Marit Bernson c/o Autoren.Services": "Bernson, Marit",
"Mark Moriani": "Moriani, Mark",
"Markus Dosch": "Dosch, Markus",
"Markus Tomshöfer": "Tomshöfer, Markus",
"Marlis Madani": "Madani, Marlis",
"Martina Wagner": "Wagner, Martina",
"Matt van Bogen": "Bogen, Matt van",
"Maximilian Devantier": "Devantier, Maximilian",
"Melissa Ratsch": "Ratsch, Melissa",
"Metz, Janos": "Metz, Janos",
"Mia Graf": "Graf, Mia",
"Mia Montiga": "Montiga, Mia",
"Micha Black c/o be-Agentur Dieter Betz": "Black, Micha",
"Michael Klüter": "Klüter, Michael",
"Michael Schönberg": "Schönberg, Michael",
"Michael Schuck": "Schuck, Michael",
"Michael Waßmann": "Waßmann, Michael",
"Mirko Perschke": "Perschke, Mirko",
"Mona-Elena Popp": "Elena, Mona-Elena",
"Monika Schweitzer": "Schweitzer, Monika",
"Monika Tietze": "Tietze, Monika",
"Müller, Otto": "Müller, Otto",
"Nadja Lanart": "Lanart, Nadja",
"Natalie J. Kohler": "Kohler, Natalie J.",
"Natalie Rabengut": "Rabengut, Natalie",
"Natascha Roth": "Roth, Natascha",
"Nick Rinnert": "Rinnert, Nick",
"Nicolas Lindt": "Lindt, Nicolas",
"Nicole Zaspel": "Zaspel, Nicole",
"Noelie Lalie": "Lalie, Noelie",
"Norbert Zajac": "Zajac, Norbert",
"Oliver Lutz": "Lutz, Oliver",
"Otger Hollerschek": "Hollerschek, Otger",
"P. Sorgenfrei": "Sorgenfrei, Paul",
"P. Urban": "Urban, Peter",
"P. Warnecke": "Warnecke, Petra",
"Peter Braukmann, edition 7eichen": "Braukmann, Peter",
"Peter Ehlert": "Ehlert, Peter",
"Peter J. Hakenjos": "Hakenjos, Peter",
"Peter Schmitt": "Schmitt, Peter",
"Peter Zörner": "Zörner, Peter",
"Petra Anna Schiener": "Schiener, Petra Anna",
"Petra Martin": "Martin, Petra",
"Phil Humor": "Humor, Phil",
"Phillip Gleichner": "Gleichner, Phillip",
"R. Ansgar Schulz-Mittenzwei": "Schulz-Mittenzwei, Ansgar",
"R. Gamperl": "Gamperl, Robert",
"R. J. Ruprecht": "Ruprecht, Rainer Joachim",
"R. Kohlase": "Kohlase, Roland",
"R. Mäder": "Mäder, Roland",
"R. Namtel": "Namtel, Rudy",
"R. Pordzik": "Pordzik, Ralph",
"R. Roos": "Roos, Renate",
"Renate Roos": "Roos, Renate",
"R. Schlegel": "Schlegel, Rolf",
"R. Stumpp": "Stumpp, Ralf",
"Rached Kaiser": "Kaiser, Rached",
"Radu Pavalache": "Pavalache, Radu",
"Raimund Kellmann": "Kellmann, Raimund",
"Ralf Boscher": "Boscher, Ralf",
"Ralph Gehrke": "Gehrke, Ralph",
"Ralph Weibel": "Weibel, Ralph",
"Regina Vogel": "Vogel, Regina",
"Reiner Kotulla": "Kotulla, Reiner",
"Reinhold Sattlegger": "Sattlegger, Reinhold",
"Renate Werner": "Werner, Renate",
"Rieger, Gisela": "Rieger, Gisela",
"Roland Reischl": "Reischl, Roland",
"Roland Springborn": "Springborn, Roland",
"Roland Wiedenmann": "Wiedenmann, Roland",
"Ronny Morgner": "Morgner, Ronny",
"Rosemarie Schnepel": "Schnepel, Rosemarie",
"Rosi Masuch": "Masuch, Rosi",
"Rudi Freiherr von Niehuus": "Niehuus, Rudi von",
"Rupp Grasegger": "Grasegger, Rupp",
"S. Dukor-Obrist": "Dukor, Steffi",
"S. Groth": "Groth, Silke",
"Silke Groth": "Groth, Silke",
"S. H. Raven": "Raven, S. H.",
"S. Köhler": "Köhler, Stefan",
"S. Piehl": "Piehl, Siegmut",
"S. Richter-Rauch": "Richter-Rauch, Sabine",
"S. Steiger": "Steiger, Siegfried",
"Sabine Kraechter": "Kraechter, Sabine",
"Sabine Welsch-Lehmann": "Welsch, Sabine",
"Said Azami": "Azami, Said",
"Sandra Andrés": "Andrés, Sandra",
"Sandra Gottwaldt": "Gottwaldt, Sandra",
"Sandra Pulletz": "Pulletz, Sandra",
"Sarah Baines c/o madera GmbH & Co. KG": "Baines, Sarah",
"Schneider, Michael, u. Nadine Schneider": "Michael, Schneider; Schneider, Nadine",
"Sebastian Wünsche": "Wünsche, Sebastian",
"Sigrid Schwab": "Schwab, Sigrid",
"Silke Heinig": "Heinig, Silke",
"Silvia Flür-Vonstadl": "Flür-Vonstadl, Silvia",
"Silvia Muster": "Muster, Silvia",
"Silvia Rohm": "Rohm, Silvia",
"Sophie Neumann": "Neumann, Sophie",
"Stefan Oliver Wlach": "Wlach, Stefan Oliver",
"Stephan Lill": "Lill, Stephan",
"Stephan Peters": "Peters, Stephan",
"Stephan Wilhelm": "Wilhelm, Stephan",
"Steven Marx": "Marx, Steven",
"Stina Milner": "Milner, Stina",
"Sybille Roth": "Roth, Sybille",
"Sylvia Kenner": "Kenner, Sylvia",
"T. Englert": "Englert, Traute",
"T. Trauf": "Trauf, T.",
"T. Wiesmann": "Wiesmann, T.",
"Tanja Hanika": "Hanika, Tanja",
"Tara Riedman c/o MedienTeam Verlag GmbH & Co KG": "Riedman, Tara",
"Thomas Karg": "Karg, Thomas",
"Thomas Martin Kirchner": "Kirchner, Thomas Martin",
"Tony Linnertz": "Linnertz, Tony",
"Torsten Hirschberger": "Hirschberger, Torsten",
"U. Mißlitz": "Misslitz, Ulrich",
"U. Mrozinski": "Mrozinski, Ute",
"U. Neubauer": "Neubauer, Ursula",
"U. Pörschke": "Pörschke, Udo",
"U. Serwuschok": "Serwuschok, Uta",
"U. Waldow": "Waldow, Ute",
"Ulrich Rach": "Rach, Ulrich",
"Uwe Klein": "Klein, Uwe",
"Uwe Kurz": "Kurz, Uwe",
"V. Wendt": "Wendt, Volker",
"Valeria Liebermann": "Liebermann, Valeria",
"Verena Sägesser Wittmer": "Sägesser Wittmer, Verena",
"Veronika Beci": "Beci, Veronika",
"Veso Portarsky": "Portarsky, Veso",
"Vincent Theodor Thomas": "Thomas, Vincent Theodor",
"Viola Plötz": "Plötz, Viola",
"Viros Amat c/o Block Services": "Amat, Viros",
"Vittoria Braun": "Braun, Vittoria",
"Volker Frehsee": "Frehsee, Volker",
"Volker Lüdecke": "Lüdecke, Volker",
"W. Loh": "Loh, Wolfgang",
"W. M. Weist": "Weist, Winfried M.",
"W. Roth": "Roth, Werner",
"W. Wurmbrand": "Wurmbrand, Walter",
"Walter Meier": "Meier, Walter",
"Walter Reick": "Reick, Walter",
"Werner Adams": "Adams, Werner",
"Werner Jurt": "Jurt, Werner",
"Werner Pfeil": "Pfeil, Werner",
"Wilfried Stubbe": "Stubbe, Wilfried",
"Wolfgang Genedl": "Genedl, Wolfgang",
"Wolfgang Görke": "Görke, Wolfgang",
"Wolfgang Martin, im Selbstverlag": "Martin, Wolfgang",
"Wolfram Jörke": "Jörke, Wolfram",
"Yvonne Kraus": "Kraus, Yvonne",
"Yvonne Taddeo": "Taddeo, Yvonne",

"Bär (ELVEA VERLAG), Michael": "Bär, Michael", 
"Bellmund Susanne Bertels Anna-Maria Böhm Martina Frank Vera Gerling Sabine Gräter Christa Hilsch, Sabine": "Bellmund, Susanne; Bertels, Anna-Maria; Böhm, Martina; Frank, Vera; Gerling, Sabine; Gräter, Christa; Hilsch, Sabine",
"Blaumanis, R?dolfs": "Blaumanis, Rūdolfs", 
"Bosse, Ay?e": "Bosse, Ays̨e", 
"Bossert, RolfFranz Hodjak": "Bossert, Rolf; Hodjak, Franz",
"brauner, Urlike": "Brauner, Ulrike",
"Bryla, Ka?ka": "Bryla, Kaśka",
"Bud?ák, Jan": "Budňák, Jan", 
"C???ma?i, R.": "Cūṭāmaṇi, R.",
"Christine Graf, Sabine Siebert, Reija Tuuli, Swantje Oppermann, Margit Heumann, Sonja Pudmensky": "Graf, Christine; Siebert, Sabine; Tuuli, Reija; Oppermann, Swantje; Heumann, Margit; Pudmensky, Sonja",
"Daniel Juhr &lt": "Juhr, Daniel", 
"Detlef W. Stein &lt": "Stein, Detlef W.",
"Dr. Bartsch, Johannes": "Bartsch, Johannes",
"Dr. Hadwiger, Heinz-Helmut": "Hadwiger, Heinz-Helmut",
"Dr. Khachapuridse, Tengis": "Khachapuridse, Tengis",
"Dr. Kiefl, Walter": "Kiefl, Walter",
"Dr. Künzler-Behncke, Rosemarie": "Künzler-Behncke, Rosemarie",
"Dr. med. Jesse, Reiner": "Jesse, Reiner",
"Dr. Minko, Sylwester": "Minko, Sylwester",
"Dr. Müller-Dum, Denise": "Müller-Dum, Denise",
"Dr. Quast, Ulrike": "Quast, Ulrike",
"Dr. Schiffter, Roland": "Schiffter, Roland",
"Dr. Starz, Hermann": "Starz, Hermann", 
"G. Mischke, Jürgen": "Mischke, Jürgen G.",
"Gehrig, Linda Hagspiel, Evi Hallermayer-Jahreiß, Elvira Kolb-Precht, Susanne Kotrus, Brigitte Mattes, Julia": "Gehrig, Julia; Hagspiel, Linda; Hallermayer-Jahreiß, Evi; Kolb-Precht, Elvira; Kotrus, Susanne; Mattes, Brigitte",
"Hörletseder, Jürgen Medium": "Hörletseder, Jürgen",
"Günter Helmes / Gernot von Blödelfels, Günter Helmes /. Gernot von": "Helmes, Günter; von Blödelfels, Gernot",
"Helga Marie Paulin, Helga Marie": "Paulin, Helga Marie",
"Helmut Haack, Helmut": "Haack, Helmut",
"Hoffmann, Heinrich Lütje, Julius": "Hoffmann, Heinrich; Lütje, Julius",
"Jochen und Renate Krohn, Jochen und Renate": "Krohn, Jochen; Krohn, Renate",
"JOREL THE CRYING GENTLEMAN, JOREL THE CRYING": "JOREL THE CRYING GENTLEMAN", 
"Julia & Marcel Preiß, Julia &. Marcel": "Preiß, Marcel; Preiß, Julia",
"Jurek P": "P., Jurek",
"Karr & Wehner, Karr &.": "Karr, Hanns-Peter; Wehner, Walther", 
"Kathrin Jarz": "Jarz, Kathrin", 
"Lindemann, Dr. phil. Thorsten": "Lindemann, Thorsten", 
"Michael G. Spitzer, Michael G.": "Spitzer, Michael G.",
"Michael Hoffmann, Michael": "Hoffmann, Michael",
"Michael, Hoffmann": "Hoffmann, Michael",
"Rotermund, Kurt Meyer-": "Meyer-Rotermund, Kurt",
"Westhäußer (Cover), Sabrina": "Westhäußer, Sabrina"
}

# Funktion zur Konsolidierung der Namen anhand des Mappings
def konsolidiere_und_speichere(file_path, columns_to_map, name_mapping, save_dir):
    df = pd.read_csv(file_path, sep=';', encoding='utf-8')

    base = os.path.basename(file_path)
    file_root, ext = os.path.splitext(base)

    for column_name in columns_to_map:
        if column_name in df.columns:
            # Ausgabe der ursprünglichen Namen zur Debugging-Zwecken
            print(f"Datei: {base}, Spalte: {column_name}")
            print("Ursprüngliche Namen: ", df[column_name].dropna().unique()[:5])

            konsolidierte_names = []

            for eintrag in df[column_name]:
                if pd.notna(eintrag):
                    names = [name.strip() for name in eintrag.split(';') if name.strip()]
                    konsolidierte_names.append(';'.join(name_mapping.get(name, name) for name in names))
                else:
                    konsolidierte_names.append(eintrag)

            df[column_name] = konsolidierte_names

            # Ausgabe der konsolidierten Namen zur Debugging-Zwecken
            print("Konsolidierte Namen: ", df[column_name].dropna().unique()[:5])

        # Neue Speicherung nach Bearbeitung der letzten Spalte
        neuer_dateiname = os.path.join(save_dir, f"{file_root.replace('_kk', '')}_kk_pk{ext}")
        df.to_csv(neuer_dateiname, index=False, encoding='utf-8', sep=';')
        print(f"✅ Datei {neuer_dateiname} erfolgreich konsolidiert und gespeichert.")

# Definiere den Speicherort für die konsolidierten Dateien
save_directory = r'C:\Pfad\zu\Konsolidierte'

# Sorge dafür, dass das Verzeichnis existiert
os.makedirs(save_directory, exist_ok=True)

# Dateipfade und zu konsolidierende Spalten
dateien = [
    # Pfad zu den Dateien und die dazugehörigen Spalten
    (r'C:\Pfad\zu\Konsolidierte\2008-2023_Buch_kk.csv', ['buchverlag', 'verfasser', 'herausgeber', 'uebersetzer', 'keinerolle', 'sonstige']),
    (r'C:\Pfad\zu\Konsolidierte\Literaturwettbewerb_kk.csv', ['veranstalter']),
    (r'C:\Pfad\zu\Konsolidierte\Literaturwettbewerb_Ausgabe_kk.csv', ['gewinner', 'jury']),
    (r'C:\Pfad\zu\Konsolidierte\Literaturzeitschrift_kk.csv', ['zeitschriftverlegtvon', 'creator']),
    (r'C:\Pfad\zu\Konsolidierte\Literaturzeitschrift_Ausgabe_kk.csv', ['verfasser_ke_beiträge']),
    (r'C:\Pfad\zu\Konsolidierte\Plattform_Internet_kk.csv', ['unternehmen'])
]

# Konsolidiere alle Dateien
for datei, spalten in dateien:
    konsolidiere_und_speichere(datei, spalten, personenname_mapping, save_directory)
