==> skripte_datenkorpus/9_Neuordnung_Koerperschaft_Person.py <==
import pandas as pd

# Listen der falsch zugeordneten Einträge
personen_in_koerperschaft = [
"Amsler, André",
"Büttner, Anne",
"Sommer, Anton F. W.", 
"Gasser, Alex", 
"Hovestadt, Alfred", 
"Langkath, Angela", 
"Thorn, Almud",
"Tschirner, Andrea",
"Wellbrock, Achim",
"Westermann, Alfons", 
"Zmyj, Annette",
"Starch, Alfred",
"Herrmann, Andrea",
"Yildiz, Andrea",
"Dölling, Andreas",
"Meier, Andreas Herbert",
"Suter, Angela",
"Horn, Angelika", 
"Bohn, Angelika",
"Kast, Anja M.",
"Redhead, Anke",
"Achner, Annegret",
"Kipnowski, Annette; Raschke, Marion",
"Eder, Anselm", 
"Meeser, Antje",
"Launer, Anton",
"Löschner, Antonia",
"Wimmer, Antonia",
"Andrä, Armin",
"Stoll, Arne", 
"Dethloff, Bettina",
"Bohmann, Birgit Evi",
"Krauß, Barbara",
"Schmid-Leuz, Barbara",
"Luipold, Bernhard",
"Pallina, B.; Milke, Natalie",
"Schwarz, Bernhard",
"Stählin-Tschanz, Betli",
"Seelk, Barbara",
"Bauer, Hilde",
"Apel, Bernd",
"Hoppe, Bernhard",
"Hellebrand, Brigitte",
"Regitz, Brigitte",
"Brigitte van Hattem",
"Linde, Brita",
"Bendixen, Britta",
"Moebius, Bruno",
"Bungert, Klauspeter",
"Vinandy, Charel Aristide",
"Anastassiou, C.",
"Gansen-Hainze, Christine",
"Greene, Carolina",
"Schartner, Christian",
"Wachsmann, Christiane",
"Raaven, C. A.",
"Mayer, Carmen",
"Gratias, Carmen Silvia",
"Heine, Carola",
"Nichte, Carsten",
"Schroer, Christopher",
"Maria, Christa",
"Schloßmacher, Christa",
"Gilberg, Christel",
"Grimm, Christian",
"Piesch, Christian",
"Seirer, Christian",
"Kromp, Christiane",
"Schwarzmüller, Christine",
"Berwert, Claudia",
"Cliff, Claudia",
"Freyer, Claudia",
"Otte, Claudia",
"Wendt, Claudia",
"Beese, Claus",
"Rivers, Cleo",
"Hartung, D.",
"Rauskolb, Doris",
"Reeken, Dieter von",
"Müller, Dana",
"Vogel, Daniela",
"Dennerlein, Irmgard",
"Brandstetter, Diana",
"Höntsch, Dieter",
"Reeken, Dieter von",
"Koch, Doro",
"Kukral, Elisabeth",
"Treml, Hubert",
"Zaffarana, Maria",
"Menke, Thorsten",
"Adler, Eckhard",
"Blaich, E.",
"Frieko, Eva",
"Grün, Erika",
"Schuster, Ewald",
"Truninger, Edith",
"Decesso, Elisabeth",
"Rüdinger, Elisabeth",
"Krys, Elke",
"Beier, Elke; Beier, Hans-Hermann",
"Hruby, Elmar S.",
"Füner, Eugen",
"Grinschuk, Eugen",
"Engel, Eva",
"Lirot, Eva",
"Klima, Eva Maria",
"Dworak, Ewald",
"Irmscher, Frank",
"Sargus, F.",
"Westermann, Felix",
"Baranowski, Frank",
"Mitschke, Frank-Thomas",
"Ucik, Fritz",
"Grob, Gisela",
"Perkams, Günter",
"Pils, Georg Hugo Donatus",
"Witt, Gerd-Dieter",
"Hofmann, Gabriela",
"Gail, Gabriele",
"Kaltofen, Gabriele",
"Oppel, Gerhard",
"Staflinger, Gerhard",
"Höhn, Gerhard W. H.",
"Rieger, Gisela",
"Schultheiss, Gudrun",
"Bruggmann, Guido",
"Catell, Gunny",
"Paffrath, Günther",
"Deinhammer, Helmut",
"Fleischle, Hans",
"Gaicher, Horst",
"Ganz, Horst",
"Glatte, Harald",
"Haßlauer, Hans",
"Heinze, Horst",
"Helnwein, Helga",
"Helnwein, Helga",
"Markert, Hans",
"Pooschke, Hilde",
"Schaper, Henning",
"Steinborn, Heinz",
"Winkelmann, Hartmut W.",
"Lorang, Hans-Peter",
"Meisel, Ulrich",
"Blum, Hajo",
"Winkler, Hannah",
"Bauer, Hans-Peter",
"Brettner, Harald",
"Grunder, Harald",
"Wiesmann, Harald",
"Thein, Heinrich",
"Draab, Hermann",
"Scherer, Heinz-Josef",
"Hartnagel, Helen",
"Demuth, Helga",
"Fleischhauer, Helmut P.",
"Boschmann, Henselowsky",
"Schnaibel, Herbert",
"Wetterauer, Herbert",
"Scheffler, Hermann",
"Hiller, Jakob",
"Hartmann, Horst",
"Rössl, Ingrid",
"Welzig, Inge",
"Stauch, Ingo",
"Heim, Ingrid",
"Feuerstein, Jenny",
"Gottanka, Josef",
"Heinze, Joachim",
"Kartheiser, Josiane",
"Reichelt, Joachim",
"Richtzenhain, Jürgen",
"Walther, Jan",
"Schön, Jes",
"Graves, Jessica",
"Bosch, Joa",
"Güntzel, Joachim",
"Gerhard, Johanna; Kempen, Sarah M.",
"Weese, Jörg",
"Mugler, Josef",
"Schmaus, Josef",
"Wieser, Josef",
"Laloi, Josephine",
"Fischer, Julia",
"Pröll, Juliane",
"Reister, Juliane",
"Berus, Jürgen",
"Hanik, Jürgen",
"Kron, J. R.",
"Ruszkowski, Jürgen",
"Schmidt, Kristina",
"Frings, Karl",
"Ganahl, Kay",
"Ganahl, Kay",
"Greve, Kurt",
"Hawkes, Karin",
"Kluge, Karlheinz",
"Pfaffeneder, Klaus",
"Sölkner, Kurt",
"Herterich, Klaus W.",
"Hirschmann, Karl-Gustav",
"Kalbhenn, Marlies",
"Hackbart, Karin",
"Peter, Karin",
"Aßmann, Karl",
"Müller, Karl",
"Rambeaud, Katharina",
"Maier, Kathrin",
"Zelenka, Ken",
"Zotzmann, Klaudia",
"Bungert, Klauspeter",
"Gruber, Kristin",
"Lieschke, Kristina",
"Neubert, Kurt F.",
"Haberstich, Kurt",
"Olsson, Lars",
"Kier, Laura",
"Andresen, Laura-Marie",
"Dierssen, Lavinia",
"Gutsche, Leon",
"Hölzle, Lisa",
"Göddemeyer, Lorenz",
"Rosenberg, Lothar",
"Böhl, Lukas",
"Pan, Luzie",
"Belz, Manfred",
"Berger, M.",
"Cane, M.",
"Frenk, Martin",
"Frenk, Martin",
"Kraus, Margit",
"Lengfelder, Martin",
"Mauthe, Margret",
"Ritter, Marion",
"Schreckenberg, Monika; Völlmicke, Kristin; Lamprecht, Michael",
"Winterhalter, Martin",
"Maar, Oliver",
"Saskia, Madame",
"Puljic, Madeleine",
"Lindner, Maik",
"Passehl, Mandy",
"Thiel, Manuela",
"Kuiper, Marcel",
"Porta, Marcel",
"Rochholz, Marco",
"Föderl, Maria",
"Meyer, Maria",
"Ocean, Marina",
"Schuster, Marina",
"Kummerow, Marion",
"Bernson, Marit",
"Moriani, Mark",
"Dosch, Markus",
"Tomshöfer, Markus",
"Madani, Marlis",
"Wagner, Martina",
"Bogen, Matt van",
"Devantier, Maximilian",
"Ratsch, Melissa",
"Metz, Janos",
"Graf, Mia",
"Montiga, Mia",
"Black, Micha",
"Klüter, Michael",
"Schönberg, Michael",
"Schuck, Michael",
"Waßmann, Michael",
"Perschke, Mirko",
"Elena, Mona-Elena",
"Schweitzer, Monika",
"Tietze, Monika",
"Müller, Otto",
"Lanart, Nadja",
"Kohler, Natalie J.",
"Rabengut, Natalie",
"Roth, Natascha",
"Rinnert, Nick",
"Lindt, Nicolas",
"Zaspel, Nicole",
"Lalie, Noelie",
"Zajac, Norbert",
"Lutz, Oliver",
"Hollerschek, Otger",
"Sorgenfrei, Paul",
"Urban, Peter",
"Warnecke, Petra",
"Braukmann, Peter",
"Ehlert, Peter",
"Hakenjos, Peter",
"Schmitt, Peter",
"Zörner, Peter",
"Schiener, Petra Anna",
"Martin, Petra",
"Humor, Phil",
"Gleichner, Phillip",
"Schulz-Mittenzwei, Ansgar",
"Gamperl, Robert",
"Ruprecht, Rainer Joachim",
"Kohlase, Roland",
"Mäder, Roland",
"Namtel, Rudy",
"Pordzik, Ralph",
"Roos, Renate",
"Roos, Renate",
"Schlegel, Rolf",
"Stumpp, Ralf",
"Kaiser, Rached",
"Pavalache, Radu",
"Kellmann, Raimund",
"Boscher, Ralf",
"Gehrke, Ralph",
"Weibel, Ralph",
"Vogel, Regina",
"Kotulla, Reiner",
"Sattlegger, Reinhold",
"Werner, Renate",
"Rieger, Gisela",
"Reischl, Roland",
"Springborn, Roland",
"Wiedenmann, Roland",
"Morgner, Ronny",
"Schnepel, Rosemarie",
"Masuch, Rosi",
"Niehuus, Rudi von",
"Grasegger, Rupp",
"Dukor, Steffi",
"Groth, Silke",
"Groth, Silke",
"Raven, S. H.",
"Köhler, Stefan",
"Piehl, Siegmut",
"Richter-Rauch, Sabine",
"Steiger, Siegfried",
"Kraechter, Sabine",
"Welsch, Sabine",
"Azami, Said",
"Andrés, Sandra",
"Gottwaldt, Sandra",
"Pulletz, Sandra",
"Baines, Sarah",
"Michael, Schneider; Schneider, Nadine",
"Wünsche, Sebastian",
"Schwab, Sigrid",
"Heinig, Silke",
"Flür-Vonstadl, Silvia",
"Muster, Silvia",
"Rohm, Silvia",
"Neumann, Sophie",
"Wlach, Stefan Oliver",
"Lill, Stephan",
"Peters, Stephan",
"Wilhelm, Stephan",
"Marx, Steven",
"Milner, Stina",
"Roth, Sybille",
"Kenner, Sylvia",
"Englert, Traute",
"Trauf, T.",
"Wiesmann, T.",
"Hanika, Tanja",
"Riedman, Tara",
"Karg, Thomas",
"Kirchner, Thomas Martin",
"Linnertz, Tony",
"Hirschberger, Torsten",
"Misslitz, Ulrich",
"Mrozinski, Ute",
"Neubauer, Ursula",
"Pörschke, Udo",
"Serwuschok, Uta",
"Waldow, Ute",
"Rach, Ulrich",
"Klein, Uwe",
"Kurz, Uwe",
"Wendt, Volker",
"Liebermann, Valeria",
"Sägesser Wittmer, Verena",
"Beci, Veronika",
"Portarsky, Veso",
"Thomas, Vincent Theodor",
"Plötz, Viola",
"Amat, Viros",
"Braun, Vittoria",
"Frehsee, Volker",
"Lüdecke, Volker",
"Loh, Wolfgang",
"Weist, Winfried M.",
"Roth, Werner",
"Wurmbrand, Walter",
"Meier, Walter",
"Reick, Walter",
"Adams, Werner",
"Jurt, Werner",
"Pfeil, Werner",
"Stubbe, Wilfried",
"Genedl, Wolfgang",
"Görke, Wolfgang",
"Martin, Wolfgang",
"Jörke, Wolfram",
"Kraus, Yvonne",
"Taddeo, Yvonne",
"Zörner, Peter", 
"Zsolnay, Paul"

]


# Liste der Körperschaften in der Person.csv 
koerperschaften_in_personen = [
"Karl-May-Verlag", '"Das Geheimnis des Überseekoffers", Autorengruppe"', "Wortspieler", "wirschreibenfürahrweiler, Autorengemeinschaft",
"Abraxas &lt", "adakia Verlag", "Aids-Hilfe NRW e.V. / Landesarbeitsgemeinschaft Frauen und Aids in NRW", 
"Akademie Ländlicher Raum Baden-Württemberg", "AKU!", "Alles wird schön e.V.", "Angst & Schrecken", 
"Arbeiterwohlfahrt Kreisverband Siegen-Wittgenstein/Olpe", 
"Arbeitskreis der Gleichstellungsbeauftragten des Märkischen Kreises, Arbeitskreis der Gleichstellungsbeauftragten des Märkischen",
"Arial-10 e.V.", "Ars Vivendi", "Ars Edition", "Authors United", "Autoren Netzwerk", 
"Autorenforum Köln e.V., Autorenforum Köln e.V.", 
"Autorenforum Phoenix &lt",
"Autorenforum Spruchreif",
"Autorengemeinschaft Myristika",
"Coburger Autorengemeinschaft Schreibsand",
"Autorengruppe Kleeblatt",
"Autorenkreis Main-Spessart",
"Autorenkreis Plesse",
"Autorenkreis Ruhr-Mark", 
"Autorentreff Bad Camberg e.V.",
"Autorinnenvereinigung e.V., Netzwerk",
"Bar-Verlag",
"Berliner Festspiele", 
"Berner Schriftstellerinnen- und Schriftsteller-Verein",
"Blaues Kreuz Schweiz",
"BördeAutoren e.V.",
"BoD - Books on Demand",
"Braumüller GmbH",
"Braunschweigische Landschaft e.V.",
"Brauseboys",
"Buchhandlung LeseZeit",
"Scripturium Buchmacherei",
"Buchstabensalat",
"Bücken & Sulzer",
"bugo Bücherei Göfis",
"Bündner Schreibbock",
"Burgwald-Verlag + Druck",
"Café Größenwahn &lt",
"Calvendo Verlag GmbH",
"Carl Stephenson Verlag",
"CCQ GmbH",
"CiTi",
"CoCon-Verlag GmbH",
"Creativo, Initiationgruppe für Literatur, Wissenschaft und KUnst",
"custos verlag",
"DeBehr",
"der VHS-Baar, Kreative Schreibwerkstatt",
"SilvForLife Design",
"Dichterstammtisch &lt",
"Die Küstenautoren",
"Die Münchner Turmschreiber e.V.",
"Die Plattform &lt",
"Die Rosenheimer Autoren",
"Die Trommel - Verein für Literatur, Harmonie & Lebensenergie",
"Dieselstrasse, Kulturzentrum",
"Diogenes-Theater",
"Docupoint-Verlag",
"Dorise Verlag",
"Dresdner Literaturner",
"e.V., Word and Shield"
"Edition 8",
"Ed. Eizenbergerhof",
"edition karo",
"Ed. lex liszt 12",
"Edition Tirol",
"Edition Winterwork",
"EDITION XXL",
"Edition, Althea S.M.",
"EDITION, S.M.",
"Edition, Trio Infernale",
"EDUTAINMENT GmbH, VISUAL",
"Eitel Kunst e.V. - Eine Sammlung UnDichter",
"Elbaol Verlag",
"Elysion Books",
"Engelsdorfer Verlag",
"EPLA Verlag",
"Epubli",
"Erich Kästner - Schreibwerkstatt",
"Evangelisches Johannesstift Berlin, Behindertenhilfe gGmbH, Die Macherei",
"F.-Coppenrath-Verlag &lt",
"Fabuloso Verlag",
"FELiCiTARiS",
"Femscript, Netzwerk Schreibender Frauen",
"Förderstudio Literatur",
"Fotolia",
"Frauenfelder Krimitage",
"Freier Deutscher Autorenverband",
"Freier Deutscher Autorenverband / Landesverband Bayern",
"Freier Deutscher Autorenverband / Landesverband Berlin",
"Freier Deutscher Autorenverband / Landesverband Hamburg-Schleswig-Holstein",
"Freier Deutscher Autorenverband / Landesverband Niedersachsen und Bremen",
"Freier Deutscher Autorenverband / Landesverband Sachsen",
"Freundeskreis, Botanischer Garten",
"Friedrich-Bödecker-Kreis in Sachsen-Anhalt",
"G. Heinrich Hofmann Inh. Jürgen Sommerer",
"Gareis Marketing",
"GEDOK e.V.",
"Geest Verlag",
"Geschichtswerkstatt Büdingen",
"Gesellschaft für neue Literatur Berlin (Hrsg.)",
"Gesellschaft Hessischer Literaturfreunde",
"Goethe-Schule Harburg",
"Goldstadt-Autoren e. V., Verein",
"Grafikdesign, MAGA",
"Groh Verlag",
"Gröls Verlag",
"Gruppe schreibender SeniorInnen Leipzig",
"guhl glaser, st.",
"Haiku-Gesellschaft, Deutsche",
"Hainburger Autorenrunde",
"Hai-tec Verlags- und Autorenservice",
"Hands Art, Four",
"Haus für Jugend Kultur und Stadtteil Stellingen",
"Haus Havelblick GmbH",
"Heimann-Stiftung, für Völkerverständigung",
"Lingen Verlag",
"Herzensmenschen",
"Hildesheim> / Arbeitsgruppe Gegen Gewalt an Schulen",
"House of Life e.V.",
"IATROS-Verlag und Services GmbH",
"ib-lyric, artfactory",
"Initiative Tannenbu(s)ch",
"Innocentia Verlag",
"Intronauten Verlag",
"Independent Bookworm",
"Jan Thorbecke Verlag",
"Jean Verlag",
"Jugendzentrum Anyway &lt",
"Jung Wien '14",
"JVA Hannover", 
"Kawohl Verlag",
"Keller + Kuhn",
"KommPlot, Autorinnengruppe",
"Kreative Schreibwerkstatt &lt",
"Kreisausschuss Odenwaldkreis, Kreisausschuss",
"Kultur im Trulliland e.V.",
"Kultur- und Tourismusverein St. Leonhard/Fr.",
"KULTURA-EXTRA, das online-magazin, KULTURA-EXTRA",
"Kulturagenten für kreative Schulen",
"Kulturstammtisch Kirchstetten",
"Künstlergruppe 14 Zoll",
"Landesheimatbund Sachsen-Anhalt",
"Lebenshilfe Berlin",
"Lesebühne Minden",
"Leseforum Oldenburg",
"Lesezeichen, Schreibzirkel",
"Letterotik",
"Lions-Club Hamburg-Moorweide / Gruppe Wort",
"Lit.Eifel e.V. und das Literaturhaus der Eifelgemeinde Nettersheim",
"Literareon",
"Literarische Gruppe Osnabrück e.V.",
"Literarischer Stammtisch Neunkirchen",
"Literarischer Verein der Pfalz",
"Literarisches Kolloquium Berlin",
"Literatennetzwerk Textur",
"Literatenrunde Karlsruhe",
"Literatur Quickie",
"Literatur Treff Grünau",
"Literatur Verein e.V., Linzgau",
"Literaturforum Karben",
"Literaturforum Schwaz",
"Literaturgruppe Wachtenburg",
"Literaturkreis Geismar",
"Literaturplattform Bucklige Welt",
"Litopian e.V., Literarischer Förderverein",
"LLC, #justfuckindoit",
"LOGO Jugendmanagement GmbH",
"Machandel Verlag",
"Marktgemeinde Winklern",
"MCE Verlag",
"Menantes-Förderkreis",
"Merquana",
"Meyer & Meyer Verlag",
"Meždunarodnyj Sojuz Nemeckoj Kul'tury",
"Mitteldeutscher Verlag",
"Mölltaler Geschichten Festival &lt",
"Net-Verlag",
"Nexx Verlag",
"NOEL-Verlag",
"Norddeutscher Rundfunk",
"Odenwaldkreis, Kreisausschuss",
"Oldenburg e.V., Leseforum",
"OsirisDruck",
"Pädagogische Hochschule Freiburg im Breisgau / Gruppe 7",
"Parzellers Buchverlag",
"Pattloch Verlag",
"Paulin?, Vita Verlag",
"Pauw Literaturmanagement",
"Pegasus Verein für Kreatives Schreiben e.V.",
"Peridexion e.V.",
"Rückenwind Pflegedienst",
"Philipp Reclam jun. Verlag",
"Pilum Literatur Verlag",
"Piper",
"Präventionsrat &lt",
"Projekt Verlag",
"Projekte Verlag Hahn",
"ProMÖLLTAL - Initiative für Bildung, Kultur, Wirtschaft und Tourismus",
"Publishing, B. L.",
"Publishing, TORRO Kids",
"Realtraum - Verein zur Förderung von Literatur, Bildender Kunst und Musik",
"Rheingauer Mundartverein e. V.",
"riva",
"Romantruhe - Buchversand",
"Rotary Club Buchholz in der Nordheide",
"Rowohlt Taschenbuch Verlag",
"Salzburger Autorengruppe",
"Salzburger Pressverein",
"Salzgitter, Print & Lettershop",
"Samenkorn e.V.",
"Sarturia-Literatur-Akademie",
"SatzZeichen, Autorengruppe",
"Schreibformat C &lt"
"Schardt Verlag",
"Schreiblust-Verlag",
"Schreibgruppe Regenbogen &lt",
"Schreibwerkstatt &lt",
"Schreibwerkstatt Wendelstein",
"Schreibwettbewerb, Lampertheimer"
"SciFan Writer's Group &lt",
"scribaria",
"Scribo Verlag",
"Senferia &lt"
'"Greifenfedern", Arbeitskreis der Senioren-Akademie Greifswalder Bodden e.V.',
"Seniorenschreibtisch der Stadt Puchheim",
"Senioren-Sport- und Kreativ-Verein Halle",
"Shutterstock",
"Sieben Verlag",
"Spaß am Lesen Verlag",
"St.-Franziskus-Gymnasium &lt"
"Stieglerhaus - Gemeinnützige Privatstiftung",
"Stiftung Rosenkreuz zur Förderung Hermetischen und Gnostischen Gedankenguts",
"stolze augen books",
"SüdOst Verlag", 
"TAK - Tiroler Autorinnen und Autoren Kooperative",
"Terre des Hommes Deutschland e.V. / Arbeitsgruppe Leipzig",
"Textfabrique51 &lt",
"Textgemeinschaft, Anthologie",
"Theater am Kirchplatz",
"Tiroler Mundartkreis",
"TösterKultur e.V.",
"Treff-Autorengruppe &lt",
"Treffen Junger Autor*innen &lt",
"tredition",
"Trierer Autoren Treff, Trierer Autoren",
"Trötsch Verlag GmbH & Co. KG",
"Ulmer Autoren Einundachtzig"
"Union Sozialer Einrichtungen &lt"
"utzverlag",
"VBG Verkehrsbetriebe Glattal AG",
"Verband Deutscher Schriftsteller / Regionalgruppe Ostbayern",
"Verband deutscher Schriftstellerinnen und Schriftsteller in ver.di – Landesverband Bayern, Verband deutscher Schriftstellerinnen und Schriftsteller in ver.di –. Landesverband",
"Verband Deutscher Schriftstellerinnen und Schriftsteller Ostbayern",
"Verband Katholischer Schriftsteller Österreichs",
"Verein Frauenhäuser Steiermark, Verein Frauenhäuser",
"Verlag almariom.de", 
"Verlag am Eschbach",
"Herder",
"Ed. Wilderermuseum",
"Verlag, Animeda",
"Chawila",
"Elvea Verlag",
"Kelebek Verlag",
"Verlag, Mondschein Corona",
"via Pho.to-Tool/ Made by Photo Lab, Fritz",
"Volk Verlag",
"Volkshochschule Pforzheim-Enzkreis",
"Kultur Referat Vorarlberg",
"VS Bayern - Verband Deutscher Schriftstellerinnen und Schriftsteller in Bayern", 
"WattLiteraten, Autorengemeinschaft",
"WBZ Wohn- und Bürozentrum", 
"WESER-KURIER, Kulturredaktion, Dr. Hendrik Werner",
"Wißner-Verlag",
"Wochen-Zeitung",
"Wolf Präsente GmbH", 
"Woll Verlag",
"WortMachereiFulda",
"Young Stories, Frankfurt", 
"Zirkel der Schreibenden &lt",
"Zytglogge Verlag",
]


# Dateipfade
person_file = r'C:\Pfad\zu\Konsolidierte\Person_kk_pk.csv'
koerperschaft_file = r'C:\Pfad\zu\Konsolidierte\Koerperschaft_kk_pk.csv'

# CSV-Dateien laden
person_df = pd.read_csv(person_file, sep=';', encoding='utf-8')
koerperschaft_df = pd.read_csv(koerperschaft_file, sep=';', encoding='utf-8')

# Verschieben von Personen von Koerperschaft -> Person
personen_zu_verschieben = koerperschaft_df[koerperschaft_df['koerperschaft_name'].isin(personen_in_koerperschaft)]
koerperschaft_df = koerperschaft_df[~koerperschaft_df['koerperschaft_name'].isin(personen_in_koerperschaft)]
personen_zu_verschieben = personen_zu_verschieben[['koerperschaft_name']].rename(columns={'koerperschaft_name': 'person_name'})

# Vermeidung von Duplikaten in Personendaten
person_df = pd.concat([person_df, personen_zu_verschieben], ignore_index=True).drop_duplicates(subset=['person_name'])

# Verschieben von Körperschaften von Person -> Koerperschaft
koerperschaften_zu_verschieben = person_df[person_df['person_name'].isin(koerperschaften_in_personen)]
person_df = person_df[~person_df['person_name'].isin(koerperschaften_in_personen)]
koerperschaften_zu_verschieben = koerperschaften_zu_verschieben[['person_name']].rename(columns={'person_name': 'koerperschaft_name'})

# Vermeidung von Duplikaten in Körperschaftsdaten
koerperschaft_df = pd.concat([koerperschaft_df, koerperschaften_zu_verschieben], ignore_index=True).drop_duplicates(subset=['koerperschaft_name'])

# Sortierung und Neuzuweisung von IDs
person_df = person_df.sort_values(by='person_name').reset_index(drop=True)
person_df['person_id'] = range(1, len(person_df) + 1)

koerperschaft_df = koerperschaft_df.sort_values(by='koerperschaft_name').reset_index(drop=True)
koerperschaft_df['koerperschaft_id'] = range(1, len(koerperschaft_df) + 1)

# Speicherung der aktualisierten Dateien mit neuem Suffix
person_file_fin = person_file.replace('_kk_pk.csv', '_kk_pk_fin.csv')
koerperschaft_file_fin = koerperschaft_file.replace('_kk_pk.csv', '_kk_pk_fin.csv')

person_df.to_csv(person_file_fin, sep=';', index=False, encoding='utf-8')
koerperschaft_df.to_csv(koerperschaft_file_fin, sep=';', index=False, encoding='utf-8')

print("✅ Daten erfolgreich übertragen und gespeichert.")
